% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{add_features_to_assay}
\alias{add_features_to_assay}
\title{Add features to an existing assay}
\usage{
add_features_to_assay(
  assay,
  slot,
  features,
  names = NULL,
  meta = NULL,
  preserve.other = FALSE
)
}
\arguments{
\item{assay}{A \linkS4class{TglowAssay}}

\item{slot}{The slot to add features to. The features are set to NA in the other slot unless preserve.other=TRUE}

\item{features}{A vector or matrix of features to append}

\item{names}{The feature names, default to colnames(features)}

\item{meta}{Feature metadata matching colnames(assay@features), each row is a feature when features is a matrix}

\item{preserve.other}{Instead of setting the other slot to NA, skip this step}
}
\value{
The assay with the extra columns on slot
}
\description{
Adds features to an existing assay
}
