% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{add_global_ids}
\alias{add_global_ids}
\title{Add a global id to a matrix of image files.
Matrix with column pattern 'ImageNumber', 'ObjectNumber' and 'Object_Number'
which will be duplicated and have a globally unique variable added.
Output of this is stored in the same column name but with suffix _Global}
\usage{
add_global_ids(matrix)
}
\description{
Add a global id to a matrix of image files.
Matrix with column pattern 'ImageNumber', 'ObjectNumber' and 'Object_Number'
which will be duplicated and have a globally unique variable added.
Output of this is stored in the same column name but with suffix _Global
}
