% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{aggregate_to_plate}
\alias{aggregate_to_plate}
\title{Convert a feature to a 96/384 plate layout by aggregation function}
\usage{
aggregate_to_plate(
  dataset,
  assay,
  slot,
  feature,
  feature.well = NULL,
  feature.plate = NULL,
  na.rm = T,
  format = "384",
  method = base::mean
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay on dataset to use}

\item{slot}{The slot to add features to. The features are set to NA in the other slot unless preserve.other=TRUE}

\item{feature}{A single feature accessible by getDataByObject}

\item{feature.well}{\linkS4class{TglowFeatureLocation} describing the well feature or NULL (takes it from datasets@feature.map),}

\item{feature.plate}{\linkS4class{TglowFeatureLocation} describing the plate feature or NULL (takes it from datasets@feature.map),}

\item{na.rm}{Should NA's be removed. Passed to method}

\item{format}{Output plate format, character either '384' or '96'}

\item{method}{A callable to use as aggregation function, mean, median, unique etc}
}
\value{
A list of plates with features aggregated
}
\description{
Convert a feature to a 96/384 plate layout by aggregation function
}
\details{
If feature is a string, method=\code{\link[base:unique]{base::unique()}}, which will work if the well only has unique values
}
