% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoy.r
\name{AnnoyBuildIndex}
\alias{AnnoyBuildIndex}
\title{Build the annoy index}
\usage{
AnnoyBuildIndex(data, metric = "euclidean", n.trees = 50)
}
\arguments{
\item{data}{Data to build the index with}

\item{metric}{Distance metric; can be one of "euclidean", "cosine", "manhattan",
"hamming"}

\item{n.trees}{More trees gives higher precision when querying}
}
\description{
This code is directly lifted from seurat's annoy implementation here:
https://github.com/satijalab/seurat/blob/master/R/clustering.R
}
\keyword{internal}
