% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoy.r
\name{AnnoySearch}
\alias{AnnoySearch}
\title{Search an Annoy approximate nearest neighbor index}
\usage{
AnnoySearch(index, query, k, search.k = -1, include.distance = TRUE)
}
\arguments{
\item{index}{Annoy index, built with AnnoyBuildIndex}

\item{query}{A set of data to be queried against the index}

\item{k}{Number of neighbors}

\item{search.k}{During the query it will inspect up to search_k nodes which
gives you a run-time tradeoff between better accuracy and speed}

\item{include.distance}{Include the corresponding distances in the result}
}
\value{
A list with 'nn.idx' (for each element in 'query', the index of the
nearest k elements in the index) and 'nn.dists' (the distances of the nearest
k elements)
}
\description{
This code is directly lifted from seurat's annoy implementation here:
https://github.com/satijalab/seurat/blob/master/R/clustering.R
}
\keyword{internal}
