% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{apply_boxcox}
\alias{apply_boxcox}
\title{Transform assay data using boxcox transform}
\usage{
apply_boxcox(
  dataset,
  assay,
  assay.out = NULL,
  trim = TRUE,
  slot = "data",
  verbose = TRUE,
  filter.iqr = F,
  ...
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay on dataset to use}

\item{assay.out}{The name of the output assay. Defaults to paste0(assay, "-trans")}

\item{trim}{Logical indicating to trim zero variance columns and NA's after applying boxcox transform}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{verbose}{Raise a warning if columns are trimmed}

\item{filter.iqr}{Remove values outside +- 2x IQR to reduce impact of outliers}

\item{...}{Arguments passed to \code{\link{boxcox_transform}}}
}
\value{
\linkS4class{TglowDataset} with new boxcox assay
}
\description{
Transform a assay using boxcox transform. If assay="image.data", result is stored in slot image.data.trans
Lambda values are added to the features data.frame: 0=log, 1=no transformation
}
\details{
Calls boxcox_transform on an assay, please read  \code{\link{boxcox_transform}} for more relvant details and options
}
