% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.r
\name{apply_image_filters}
\alias{apply_image_filters}
\title{Filter a TglowDataset at the image level}
\usage{
apply_image_filters(dataset, filter.res)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{filter.res}{A logical matrix or vector indicating which images to keep}
}
\value{
The filtered \linkS4class{TglowDataset}
}
\description{
Apply image level filter to a TglowDataset. All objects in that image are
removed. If filter.res is a matrix, all rows must be true to keep the image.
}
