% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{boxcox_transform}
\alias{boxcox_transform}
\title{Transform assay data using boxcox transform}
\usage{
boxcox_transform(
  x,
  return.lambda = FALSE,
  mode = "boxcox",
  limit = 5,
  fudge = 0.2,
  downsample = NULL,
  add.lambda = FALSE,
  filter.iqr = FALSE
)
}
\arguments{
\item{x}{A numeric vector with values to transform}

\item{return.lambda}{return the lambda value instead}

\item{mode}{Either 'boxcox' or 'modulus', see details}

\item{limit}{The absolute value of lambda's to test}

\item{fudge}{Fudgefactor to apply either log transform, or just leave it as is}

\item{downsample}{Calculate labmda's on a random sample of the data. Either NULL (no downsampling), an integer (selects x values), or a selection vector}

\item{add.lambda}{Return a list with two items instead, one transformed x, two, the lamda used}

\item{filter.iqr}{Should values outside the 2xIQR be removed prior to estimating lambda}
}
\value{
A numeric vector with the boxcox transformed data or the optimal lambda value
}
\description{
Transform a vector using boxcox transform
If value is not positive it is offset to make everything positive
}
\details{
In mode boxcox \code{\link[MASS:boxcox]{MASS::boxcox()}} is fit on the offset values if there are any values < 0. Values are made positive by adding the min absolute value
In mode modulus the modulus transform is applied instead, which is a generalization of boxcox. See \code{\link[=modulus]{modulus()}}
}
