% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.r
\name{calculate.feature.filters}
\alias{calculate.feature.filters}
\title{Find which feature filters to a tglow dataset}
\usage{
calculate.feature.filters(dataset, filters, assay, slot, features = NULL)
}
\arguments{
\item{dataset}{A tglow dataset}

\item{filters}{A list with TglowFilter objects to apply}

\item{assay}{The assay to use for calculating filters}

\item{slot}{The slot to use for calculating filters: "data" or "scale.data"}

\item{features}{An optional subset of features to use for calculation.}
}
\value{
A logical matrix of nrow(dataset) x length(filters) where T indicates filter pass and F indicates filter fail.
}
\description{
All filters are inclusive
}
