% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.r
\name{calculate_feature_filters}
\alias{calculate_feature_filters}
\title{Calculate feature filter results for a TglowDataset}
\usage{
calculate_feature_filters(
  dataset,
  filters,
  assay,
  slot,
  features = NULL,
  na.fail = TRUE
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{filters}{A list with TglowFilter objects to apply}

\item{assay}{The assay on dataset to use}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{features}{An optional subset of features to use for calculation}

\item{na.fail}{Should NA be treated as fail, defaults to yes}
}
\value{
A logical matrix of ncol(dataset[\link{assay}]) x length(filters) where T indicates filter pass and F indicates filter fail
}
\description{
Calculate a feature filter table on a \linkS4class{TglowDataset}
All filters are inclusive
}
