% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_dataset_assay_slot}
\alias{check_dataset_assay_slot}
\title{Check a dataset, assay, slot}
\usage{
check_dataset_assay_slot(dataset, assay, slot, assay.image = NULL)
}
\description{
Check a dataset, assay slot for validity or throw an error
}
\keyword{internal}
