% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_filter_list}
\alias{check_filter_list}
\title{Check a list of TglowFilters for validitiy}
\usage{
check_filter_list(filters)
}
\description{
Check a list of TglowFilters for validitiy
}
\keyword{internal}
