% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{check_unused_covar}
\alias{check_unused_covar}
\title{Check covariates.dont.use and expand factor levels}
\usage{
check_unused_covar(data, covariates.dont.use)
}
\description{
Check covariates.dont.use and expand factor levels
}
\keyword{internal}
