% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{default_feature_map_cellprofiler}
\alias{default_feature_map_cellprofiler}
\title{Default feature map for CellProfiler parquet output}
\usage{
default_feature_map_cellprofiler()
}
\value{
A \linkS4class{TglowFeatureMap}
}
\description{
Returns the \linkS4class{TglowFeatureMap} matching the output of \code{\link{read_cellprofiler_parquet}}.
x/y/z are taken from \verb{@meta}, plate, well and field from \verb{@image.meta}.
}
