% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{default_feature_map_pipeline}
\alias{default_feature_map_pipeline}
\title{Feature map for tglow-pipeline intensity output}
\usage{
default_feature_map_pipeline(version = "latest")
}
\arguments{
\item{version}{Pipeline version, or "latest" for the most recent supported version}
}
\value{
A \linkS4class{TglowFeatureMap}
}
\description{
Returns the \linkS4class{TglowFeatureMap} matching the output of \code{\link{read_pipeline_parquet}}
for a given tglow-pipeline version (manifest.version in the pipeline's nextflow.config).
}
