% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{fast_colscale}
\alias{fast_colscale}
\title{Faster alternative to scale a matrix}
\usage{
fast_colscale(
  x,
  center = TRUE,
  scale = TRUE,
  add_attr = TRUE,
  na.rm = F,
  scale.method = "mean",
  reference.group = NULL,
  na.zero = F
)
}
\arguments{
\item{x}{A matrix}

\item{center}{Should data be centered}

\item{scale}{Should data be scaled}

\item{add_attr}{Should means and variances be added to output}

\item{na.rm}{Should NA's be removed during mean and variance calculations}

\item{scale.method}{Scaling method, either \code{mean} for z-score or \code{median} modified z-score}

\item{reference.group}{Calculate means/medians, sds/mads only on these samples but apply it to all}

\item{na.zero}{Set values that are exactly 0 to NA when calculating means and variances. See details}
}
\value{
The scaled matrix
}
\description{
Adapted from https://www.r-bloggers.com/2016/02/a-faster-scale-function/
Yields a 10-20 percent speedup and adds modified z-score scaling
}
\details{
\code{scale.method - mean}

Standard centering/scaling: (x-mean(x)) / sd(x)

\code{scale.method - median}

Median adjusted deviation centering/scaling: (x-median(x)) / mad(x)

\code{na.zero}
This removes values which are exactly zero from mean and variance calculations. Some features
might have a lot of zeroes, which can skew the mean and variance estimates and give issues when
scaling to reference samples. This sets them to NA when calculating the scaling factors.
}
\examples{
# Generate a matrix, increase the size to see a bigger speed difference
r <- matrix(runif(10000 * 100), nrow = 10000, ncol = 100)

# Run using scale
system.time({
    r.scale <- scale(r)
})

# Run using fast_colscale
system.time({
    r.fast_colscale <- fast_colscale(r, add_attr = FALSE)
})

# Compare results, ignoring the scaled:center and scaled:scale attributes set by scale()
all.equal(r.scale, r.fast_colscale, check.attributes = FALSE)
}
