% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_rep_objects.r
\name{fetch_representative_object}
\alias{fetch_representative_object}
\title{Retrieve a cell (and its neighbours) based on a feature sumstat}
\usage{
fetch_representative_object(
  dataset,
  assay,
  slot,
  feature,
  metric = "mean",
  na.rm = F,
  n = 0,
  subset = NULL,
  q = NULL
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay on dataset to use}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{feature}{The feature to find a representative objects for}

\item{metric}{Can be 'mean', 'median', 'upper.q', 'lower.q'}

\item{na.rm}{Should NA's be removed}

\item{n}{How many objects either side of the representative object should be returned}

\item{subset}{Look only in a subset of objects. Must be a selection vector}

\item{q}{Override the quantile when using upper.q (0.75) or lower.q (0.25)}
}
\value{
vector of indices in objects matrix
}
\description{
Gets a objects and its closes neighbours based on a single feature and a
sumstat (mean, median, upper.q, lower.q). To customise the quantile used
specify q.
}
