% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.r
\name{find_outliers_pca_fixed}
\alias{find_outliers_pca_fixed}
\title{Find outliers in PCA space}
\usage{
find_outliers_pca_fixed(
  dataset,
  assay,
  thresh = 3.5,
  qc.group = NULL,
  pc.n = 5,
  slot = "data",
  method = "z",
  return.pcs = FALSE,
  use_irlba = TRUE
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay to use}

\item{thresh}{Threshold in absolute PC to consider an outlier}

\item{qc.group}{A vector indicating if PC's should be calculated in subgroups of the data. Default find outliers using all objects at once}

\item{pc.n}{The number of PC's to use}

\item{method}{Method to scale PC's prior to selecting thresh. Value can be 'z' for z-score or 'mod.z' for modified zscore}

\item{return.pcs}{Should the grouped PC's be returned?}

\item{use_irlba}{Logical if \code{\link[=prcomp_irlba]{irlba::prcomp_irlba()}} or \code{\link[=prcomp]{base::prcomp()}} should be used for PCA}
}
\value{
Logical indicating if object is an outlier in pca space, or a list if return.pcs=TRUE
}
\description{
Find outliers in the PCA space of a TglowAssay. Works for both
assays and image.data. This is a simple implementation based on a fixed number
of components. For A more flexible implementation see \code{\link{find_outliers_pca}}
}
\details{
To run on image data, just specify assay="image.data"|"image.data.trans"|"image.data.norm" which is implemented
as a special case
This does not use existing dimension reductions to avoid issues when running with different QC groups
}
