% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get.feature.meta.from.names}
\alias{get.feature.meta.from.names}
\title{Construct a feature metadata from cellprofiler feature names}
\usage{
get.feature.meta.from.names(feature.names)
}
\arguments{
\item{featue.names}{Character vector of feature names}
}
\value{
A data.frame with feature attributes
}
\description{
Constructs a table extracting attributes from the names in the from
<object>_<category>_<feature-name>
}
