% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_feature_meta_from_names}
\alias{get_feature_meta_from_names}
\title{Construct a feature metadata from cellprofiler feature names}
\usage{
get_feature_meta_from_names(feature.names)
}
\arguments{
\item{feature.names}{Character vector of feature names}
}
\value{
A data.frame with feature attributes
}
\description{
Constructs a table extracting attributes from the names in the from
\if{html}{\out{<object>}}\emph{\if{html}{\out{<category>}}}\if{html}{\out{<feature-name>}}
}
