% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{get_feature_meta_from_names_pipeline}
\alias{get_feature_meta_from_names_pipeline}
\title{Get feature metadata from tglow-pipeline feature names}
\usage{
get_feature_meta_from_names_pipeline(feature.names)
}
\arguments{
\item{feature.names}{Character vector of feature names}
}
\value{
A data.frame with columns id, object, measurement, category and name
}
\description{
Counterpart of \code{\link{get_feature_meta_from_names}} for tglow-pipeline names of the
form \verb{ch\{N\}__\{stat\}}. Names without '__' (area, n_nuclei) are assigned to object 'cell'.
}
