% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tglow_dataset.r
\name{getDataByObject,TglowDataset-method}
\alias{getDataByObject,TglowDataset-method}
\title{Get image data and features per object (cell)}
\usage{
\S4method{getDataByObject}{TglowDataset}(
  object,
  j,
  assay = NULL,
  assay.image = NULL,
  slot = "data",
  drop = TRUE
)
}
\arguments{
\item{object}{A \linkS4class{TglowDataset}}

\item{j}{Character with column names from assay, assay.image, image.meta to select}

\item{assay}{The assay to select from}

\item{assay.image}{Which image assay to use, "image.data", "image.data.trans" or "image.data.norm"}

\item{slot}{Slot to fetch features from: "data" or "scale.data"}

\item{drop}{Should cols be dropped or not}
}
\value{
A data frame with the corresponding columns
}
\description{
Select columns from assay, assay.image, image.meta from 'data' or 'scale.data' slots
and return them as a data.frame
}
