% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{getDataByObject}
\alias{getDataByObject}
\title{Get image data and features per object (cell)}
\usage{
getDataByObject(
  dataset,
  j,
  assay = NULL,
  assay.image = NULL,
  slot = "data",
  drop = TRUE
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{j}{Character with column names from assay, assay.image, image.meta to select}

\item{assay}{The assay to select from. If not fetching assay columns, leave at NULL}

\item{assay.image}{Which image assay to use, "image.data", "image.data.trans" or "image.data.norm". If not fetching image.data columns, leave at NULL}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{drop}{Should cols be dropped or not}
}
\value{
A data frame with the corresponding columns
}
\description{
Select columns from assay, assay.image, image.meta from 'data' or 'scale.data' slots
and return them as a data.frame
}
