% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tglow_dataset.r
\name{getImageDataAndFeatures,TglowDataset-method}
\alias{getImageDataAndFeatures,TglowDataset-method}
\title{Get image data and features per cell}
\usage{
\S4method{getImageDataAndFeatures}{TglowDataset}(object, j, assay, slot = "data", drop = F)
}
\arguments{
\item{object}{TglowDataset}

\item{j}{character with column names from image.meta, image.data or assay@slot to select}

\item{assay}{the assay to select from}

\item{slot}{slot to fetch features fromdata or scale.data}

\item{drop}{should cols be dropped or not}
}
\value{
A data frame with the corresponding columns
}
\description{
Get image data and features per cell
}
