% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tglow_dataset.r
\name{getImageDataByObject,TglowDataset-method}
\alias{getImageDataByObject,TglowDataset-method}
\title{Fetch image data or meta data from a tglow object per object (cell).}
\usage{
\S4method{getImageDataByObject}{TglowDataset}(object, j, assay.image, slot = "data", drop = TRUE)
}
\arguments{
\item{object}{A \linkS4class{TglowDataset}}

\item{j}{Character with column names from image.meta or assay.image to select}

\item{assay.image}{Which image assay to use, "image.data", "image.data.trans" or "image.data.norm"}

\item{slot}{Slot to fetch features from data or scale.data}

\item{drop}{Should cols be dropped or not}
}
\value{
A data frame with the corresponding columns
}
\description{
Select columns from assay.image, image.meta and return them
as a data.frame per object (cell).
}
