% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{getImageDataByObject}
\alias{getImageDataByObject}
\title{Fetch image data or meta data from a tglow object per object (cell)}
\usage{
getImageDataByObject(
  dataset,
  j,
  assay.image = NULL,
  slot = "data",
  drop = TRUE
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{j}{Character with column names from image.meta or assay.image to select}

\item{assay.image}{Which image assay to use, "image.data", "image.data.trans" or "image.data.norm". If not fetching image.data columns, leave at NULL}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{drop}{Should cols be dropped or not}
}
\value{
A data frame with the corresponding columns
}
\description{
Select columns from assay.image, image.meta and return them
as a data.frame per object (cell)
}
