% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{hex_to_rgb}
\alias{hex_to_rgb}
\title{Convert a hex code to RGB}
\usage{
hex_to_rgb(hex)
}
\arguments{
\item{hex}{String with hex code}
}
\value{
a vector of length 3 with relative scaling between RGB channels between 0 and 1
}
\description{
Convert a hex code to RGB
}
