% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{imageIds}
\alias{imageIds}
\alias{imageIds<-}
\title{Get and set image Ids}
\usage{
imageIds(object)

imageIds(object) <- value
}
\arguments{
\item{object}{The object to get or set}

\item{value}{If setting using <- the new object ids to assign}
}
\value{
The object with the new Ids
}
\description{
Get and set image Id's on TglowDataset
}
