% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{img_apply_color}
\alias{img_apply_color}
\title{Apply a relative scaling of RGB values to a 3d array}
\usage{
img_apply_color(image, rgb)
}
\arguments{
\item{image}{3d array with 3rd dimension as rgb}

\item{rgb}{vector of length 3 with scaling factors for rgb channels}
}
\value{
The scaled image
}
\description{
3rd dimension is assumed to contain the color channels.
If input is grey, returns the image in the color
}
