% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{img_composite}
\alias{img_composite}
\title{Combine a series of RGB images into a composite image}
\usage{
img_composite(images, colors)
}
\arguments{
\item{images}{list of RGB EBImages}

\item{colors}{list of hex codes to color shift, must be same length as images}
}
\value{
A single RGB images representing the average color values
}
\description{
Apply a color shift and average their RGB values at each pixel to create a composite image
}
