% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{img_max_per_channel}
\alias{img_max_per_channel}
\title{Find the maximum value in each channel in a collection of images
This is used for normalizing images}
\usage{
img_max_per_channel(images, channel.dim = 3, q = 1)
}
\arguments{
\item{images}{A list of EBImage objects, with 3 or 4 dimensions. 3d dimension
is assumed to be channel}

\item{channel.dim}{Dimension in image stack that is channel}

\item{q}{The quantile to normalize to}
}
\value{
A vector with the max values in each channel

importFrom EBImage imageData
}
\description{
Find the maximum value in each channel in a collection of images
This is used for normalizing images
}
