% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{img_max_project}
\alias{img_max_project}
\title{Max project a EBImage along the 4th axis}
\usage{
img_max_project(img)
}
\arguments{
\item{img}{EBImage with 4 dimensions}
}
\value{
An EBImage with the 4th dimension collapsed by taking the maximum value
}
\description{
Collapse the 4th dimension (z-stack or time) of an EBImage by
taking the maximum value at each pixel position across all planes.
}
