% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{img_pad_center}
\alias{img_pad_center}
\title{Pad an array by zero-padding the first two dimensions with centering}
\usage{
img_pad_center(input, target.rows, target.cols)
}
\arguments{
\item{input}{An array or matrix of any dimensionality.}

\item{target.rows}{Integer. The desired number of rows in the output (first dimension).}

\item{target.cols}{Integer. The desired number of columns in the output (second dimension).}
}
\value{
An array with the first two dimensions padded or cropped to the target size,
with the input array centered. Other dimensions remain the same as the input.
}
\description{
This function takes an input array or matrix of any number of dimensions and
pads or crops it along the first two dimensions to a specified target size.
The input is centered in the output array, and zeros are used for padding.
Dimensions beyond the first two are preserved.
If the input is larger than the target size, it is cropped accordingly.
}
\examples{
input <- array(1, dim = c(5, 5, 3))
padded <- img_pad_center(input, 10, 10)
dim(padded)  # 10 10 3
print(padded[,,1])  # Shows the first 2D slice padded and centered

}
