% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{img_to_base64png}
\alias{img_to_base64png}
\title{Convert a matrix to a base64 text PNG
Takes a 2d (greyscale) or 3d (RGB) matrix and converts it to a base64 encoded png}
\usage{
img_to_base64png(mat, add.html = F, width = 250)
}
\arguments{
\item{mat}{A 2d or 3d matrix}

\item{add.html}{Should a HTML tag be added}

\item{width}{If add.html, what should the width be of the \if{html}{\out{<img>}} tag}
}
\value{
a base64 encoded version of the image, or HTML \if{html}{\out{<img>}} tag with the data
}
\description{
Convert a matrix to a base64 text PNG
Takes a 2d (greyscale) or 3d (RGB) matrix and converts it to a base64 encoded png
}
