% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{isAvailable}
\alias{isAvailable}
\title{Check if column names are available on a dataset}
\usage{
isAvailable(dataset, j, assay, assay.image = NULL, slot, return.names = FALSE)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{j}{Character with column names from assay, assay.image, image.meta to select}

\item{assay}{The assay to select from. If not fetching assay columns, leave at NULL}

\item{assay.image}{Which image assay to use, "image.data", "image.data.trans" or "image.data.norm". If not fetching image.data columns, leave at NULL}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{return.names}{If TRUE, return a character vector of available names instead of a logical}

\item{drop}{Should cols be dropped or not}
}
\value{
A logical (or named logical vector if return.names=TRUE) indicating column availability
}
\description{
Check whether column names are available on the dataset
}
