% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{isValid}
\alias{isValid}
\title{Check if a Tglow family of objects is valid}
\usage{
isValid(object, object.names = NULL)
}
\arguments{
\item{object}{A \linkS4class{TglowDataset}, \linkS4class{TglowAssay}, \linkS4class{TglowReduction} or \linkS4class{TglowMatrix}}

\item{object.names}{A character vector of object names to validate against. If NULL, uses object@object.ids}
}
\value{
A logical indicating validitiy
}
\description{
Greedily check if a Tglow family of objects is valid.
Works on TglowDataset, TglowAssay, TglowReduction, TglowMatrix
}
\details{
Checks are peformed greedily, so as soon as an issue if found, FALSE is returned
and a warning for that issue is raised with more details. I.e. finding one issue
does not mean there are not other issues!

For TglowDataset, object.names can be left at NULL, as it uses @object.ids slot to compare against.
If not NULL it will use the vector of names you supply to check

TglowAssay, TglowReduction and TglowMatrix must all have object.names supplied, otherwise an error
is thrown.
}
