% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{list_has_overlap}
\alias{list_has_overlap}
\title{Check if there is any overlap within a list of vectors}
\usage{
list_has_overlap(list)
}
\arguments{
\item{list}{A list of vectors to overlap}
}
\value{
logical indicating if any items overlap in the list
}
\description{
Check if there is any overlap within a list of vectors
}
