% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{lmm_matrix}
\alias{lmm_matrix}
\title{Calculate linear coefficients using a Linear mixed model}
\usage{
lmm_matrix(
  response,
  design,
  formula,
  formula.null = NULL,
  residuals.only = FALSE,
  return.residuals = FALSE,
  refit = TRUE,
  control = lme4::lmerControl(check.conv.singular = list(action = "ignore", tol = 1e-04)),
  ...
)
}
\arguments{
\item{response}{A matrix with response variables}

\item{design}{The matrix with predictor variables}

\item{formula}{The latter component of the formula as a string. I.e. \code{~ x + (1|donor)}}

\item{formula.null}{The null formula for a LRT as a string. The latter component of the formula. I.e. \code{~ x + (1|donor)}}

\item{residuals.only}{Only return the residual matrix}

\item{return.residuals}{Return residual matrix in the output list. Defaults to T if residual.only = TRUE}

\item{refit}{Refit the models using MLE during the LRT (anova)}

\item{control}{Control parameters passed to lmerTest::lmerControl(). Set to ignore singularity by default.}

\item{...}{Remaining parameters passed to \code{\link[lmerTest:lmer]{lmerTest::lmer()}}}
}
\value{
A list with regression results
}
\description{
Fit a linear mixed model using REML / MLE (lme4) and find coefficients
}
\details{
By default, singularity warnings are ignored in the lmer fitting as these are common with complex designs and some HCI features.
Please check the singularity status in the output model.stats 'singular_reff' and 'singular_reff_null'. If you have many singular fits,
its advised to simplify your model as your random effect sructure might be too complex for the data at hand.
}
