% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{make_ident_labels}
\alias{make_ident_labels}
\title{Create a matrix with labels on ident positions
This function is used to get the center position of an ident in XY space
This is used to plot the cluster numbers on the UMAP plot}
\usage{
make_ident_labels(x, y, memberships)
}
\arguments{
\item{x}{The x positon}

\item{y}{The y position}

\item{vector}{of ident memberships}
}
\value{
A matrix with xy positions and a label
}
\description{
Create a matrix with labels on ident positions
This function is used to get the center position of an ident in XY space
This is used to plot the cluster numbers on the UMAP plot
}
