% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{merge.filesets}
\alias{merge.filesets}
\title{Merge list of filesets into one}
\usage{
\method{merge}{filesets}(data)
}
\arguments{
\item{data}{List output from \code{\link{tglow.read.fileset}}}
}
\value{
The mered output from \code{\link{tglow.read.fileset}}
}
\description{
Data must be a list of lists with outputs from tglow.read.fileset.a/b
or have the items, cells, meta, orl,  [children], [features], [cells_norm]
}
