% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{merge_filesets}
\alias{merge_filesets}
\title{Merge list of filesets into one}
\usage{
merge_filesets(data, skip.orl = FALSE)
}
\arguments{
\item{data}{List output from \code{\link{read_cellprofiler_fileset_a}} or \code{\link{read_cellprofiler_fileset_b}}}

\item{skip.orl}{Should object relationships be skipped}
}
\value{
The merged output from \code{\link{read_cellprofiler_fileset_a}} or \code{\link{read_cellprofiler_fileset_b}}
}
\description{
Data must be a list of lists with outputs from tglow.read.fileset.a/b
or have the items, cells, meta, orl,  \link{children}, \link{features}, \link{cells_norm}
}
