% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{modul}
\alias{modul}
\alias{modul_std}
\title{Modulus transformation}
\usage{
modul(y, lambda = lambda)

modul_std(y, lambda)
}
\arguments{
\item{y}{Numeric vector to transform}

\item{lambda}{Transformation parameter lambda}
}
\value{
The transformed numeric vector
}
\description{
As trafo's implementation of the modulus (signed power) transform.
https://github.com/cran/trafo/blob/master/R/trafos.R
}
