% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{modulus}
\alias{modulus}
\title{Transform vector using modulus transform, finding optimal lambda}
\usage{
modulus(
  object,
  lambda = seq(-2, 2, 1/10),
  plotit = TRUE,
  interp = (plotit && (m < 100)),
  eps = 1/50,
  xlab = expression(lambda),
  ylab = "log-Likelihood",
  ...
)
}
\arguments{
\item{object}{A fitted lm object or formula passed to lm()}

\item{lambda}{Sequence of lambda values to evaluate (default seq(-2, 2, 1/10))}

\item{plotit}{Should a plot of log-likelihood vs lambda be produced?}

\item{interp}{Should spline interpolation be used for the plot curve?}

\item{eps}{Tolerance for treating lambda as zero}

\item{xlab}{X-axis label for the lambda plot}

\item{ylab}{Y-axis label for the lambda plot}
}
\value{
The optimal lambda and, if plotit=TRUE, a plot of log-likelihood vs lambda
}
\description{
Modelled after MASS:boxcox() and shares the same basic signature
So see \code{\link[=MASS:boxcox]{MASS:boxcox()}} for help
}
\details{
Implements modulus transformation described here:
https://scales.r-lib.org/reference/boxcox_trans.html
John, J. A., & Draper, N. R. (1980). An alternative family of transformations. Applied Statistics, 190-197. http://www.jstor.org/stable/2986305
}
