% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nearest_index}
\alias{nearest_index}
\title{Find nearest index}
\usage{
nearest_index(x, value)
}
\arguments{
\item{x}{A vector X}

\item{value}{The value to which X should be closest}
}
\value{
The position in X where value is closest to
}
\description{
Find the index of the value in x that is closest to value
}
