% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_box}
\alias{plot_box}
\title{Plot boxplot with a numerically ordered x and a loess line}
\usage{
plot_box(
  x,
  y,
  xlab = "x",
  ylab = "y",
  main = "",
  facet = NULL,
  facet.ncol = NULL
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{main}{title}

\item{facet}{Vector with facet variable}

\item{facet.ncol}{Number of columns in facet}

\item{levels}{The order of the levels on the x-axis}
}
\value{
A ggplot2 object
}
\description{
Plot boxplot with a numerically ordered x and a loess line
}
