% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_boxline}
\alias{plot_boxline}
\title{Plot boxplot with a numerically ordered x and a loess line}
\usage{
plot_boxline(
  x,
  y,
  levels,
  xlab = "x",
  ylab = "y",
  main = "",
  line.col = "#3d403d",
  facet = NULL,
  facet.ncol = NULL,
  method = NULL
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{levels}{The order of the levels on the x-axis}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{main}{title}

\item{line.col}{Color of the line}

\item{facet}{Vector with facet variable}

\item{facet.ncol}{Number of columns in facet}

\item{method}{Method for fitting the line. Anything supported by \code{\link{ggplot2::geom_smooth()}}}
}
\value{
A ggplot2 object
}
\description{
Plot boxplot with a numerically ordered x and a loess line
}
