% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_hex}
\alias{plot_hex}
\title{Plot a hexbin scatterplot colored on density}
\usage{
plot_hex(
  x,
  y,
  bins = 50,
  binwidth = NULL,
  do.lm = T,
  lm.col = "lightgrey",
  xlab = "x",
  ylab = "y",
  main.prefix = NULL,
  main = NULL,
  facet = NULL,
  ...
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{bins}{bins}

\item{do.lm}{Should a linear fit be rendered}

\item{lm.col}{Color of the line}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{main}{title}

\item{facet}{Vector to facet on}

\item{...}{Arguments to facet_wrap}
}
\value{
A ggplot2 object
}
\description{
Plot a hexbin scatterplot colored on density
}
