% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_hist}
\alias{plot_hist}
\title{Plot a grouped histogram}
\usage{
plot_hist(
  x,
  z = NULL,
  xlab = "Value",
  main = NULL,
  bins = 100,
  facet = NULL,
  facet.ncol = NULL,
  density = FALSE,
  add.median = TRUE
)
}
\arguments{
\item{x}{x}

\item{z}{grouping variable}

\item{xlab}{xlab}

\item{main}{title}

\item{bins}{bins}

\item{facet}{Vector with facet variable}

\item{facet.ncol}{Number of columns in facet}

\item{density}{Plot as a density plot instead}

\item{add.median}{Add a vline for median}
}
\value{
A ggplot2 object
}
\description{
Plot a grouped histogram
}
