% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_img}
\alias{plot_img}
\title{Plot an individual EBImage image with a title}
\usage{
plot_img(
  img,
  main = "img",
  marker.add = F,
  marker.x = NULL,
  marker.y = NULL,
  marker.col = "white",
  marker.size = 8,
  marker.shape = 1,
  scale.bar.size = 0,
  scale.bar.pixel.size = 0.149,
  scale.bar.padding = 5,
  scale.bar.location = "bottomright",
  scale.bar.labelsize = 3
)
}
\arguments{
\item{img}{The image}

\item{main}{title}

\item{marker.add}{Should a marker be added at the center of the image}

\item{marker.x}{x position of marker. If NULL = center}

\item{marker.y}{y position of marker. If NULL = center}

\item{marker.col}{color of the marker}

\item{marker.size}{size of marker}

\item{marker.shape}{shape of marker}

\item{scale.bar.size}{Size of the scale bar in microns. Set to 0 to disable}

\item{scale.bar.pixel.size}{Size of a pixel in microns}

\item{scale.bar.padding}{Padding of the scale bar from the image border in pixels}

\item{scale.bar.location}{Location of the scale bar, one of "bottomright", "bottomleft", "topright", "topleft"}

\item{scale.bar.labelsize}{Size of the scale bar label text. Set to 0 to disable}
}
\value{
A ggplot2 object
}
\description{
Plot an individual EBImage image with a title
}
