% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_img_set}
\alias{plot_img_set}
\title{Plot a series of EBImage images with a title}
\usage{
plot_img_set(
  imgs,
  ncol,
  main = "",
  main.sub = NULL,
  text.col = "white",
  background.col = "black",
  byrow = F,
  ...
)
}
\arguments{
\item{imgs}{List of images}

\item{ncol}{Numer of columns in grid}

\item{main}{Main plot title}

\item{main.sub}{Vector of length(imgs) indicating the subplot titles}

\item{text.col}{Color of the text}

\item{background.col}{Color of the background}

\item{byrow}{Should plots be filled per row or per column}

\item{...}{Remaining arguments passed to \code{\link{plot_img}}}
}
\value{
A ggplot2 object
}
\description{
Plot a series of EBImage images with a title
}
