% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_simple_hm}
\alias{plot_simple_hm}
\title{Simple heatmap with auto labels}
\usage{
plot_simple_hm(
  data,
  cellsize = -1,
  cellwidth = 12,
  cellheight = 12,
  limits = NULL,
  cluster = T,
  range = "symmetric",
  palette = NULL,
  border = NA,
  ...
)
}
\arguments{
\item{data}{A matrix}

\item{cellsize}{The size of the cells in x and y}

\item{cellwidth}{The size of the cells in x}

\item{cellheight}{THe size of the cells in y}

\item{limits}{Clip the data to upper and lower, must be a numeric of length 2}

\item{cluster}{Should clustering be applied}

\item{range}{Automatic setting for color range: 'symmetric', 'absolute', 'auto'}

\item{palette}{Palette to use in \code{\link{RColorBrewer::brewer.pal()}}}

\item{border}{Border color}

\item{...}{Remaining arguments passed to  \code{\link{pheatmap::pheatmap()}}}
}
\value{
The pheatmap plot
}
\description{
Simple heatmap with auto labels
}
