% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_violin}
\alias{plot_violin}
\title{Plot violin plot with a density histogram to the right}
\usage{
plot_violin(
  x,
  y,
  violin = T,
  facet = NULL,
  q = c(0.05, 0.5, 0.95),
  main = NULL,
  xlab = "x",
  ylab = "y",
  stat.size = 4,
  plot.density = FALSE,
  density.widths = c(3, 1)
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{violin}{Should violin plot be added}

\item{facet}{Vector with facet variable}

\item{q}{Quantiles to add as horizontal lines in the density plot}

\item{main}{title}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{stat.size}{Size of the statistics points/lines}

\item{plot.density}{Should a density plot be added to the right of the violin/boxplot}

\item{density.widths}{Widths of the violin/boxplot and density plot when plot.density=TRUE}
}
\value{
A ggplot2 object
}
\description{
Plot violin plot with a density histogram to the right
}
