% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_xy}
\alias{plot_xy}
\title{Scatterplot using ggplot}
\usage{
plot_xy(
  x,
  y,
  xlab = "X",
  ylab = "Y",
  main = NA,
  main.prefix = "",
  size = 1,
  col = "black",
  fixed = F,
  alpha = 0.75,
  shape = 16,
  lm.col = "blue",
  do.lm = T,
  method = "lm",
  lm.group = NULL,
  raster = F,
  dpi = 300,
  facet = NULL,
  facet.ncol = NULL
)
}
\arguments{
\item{x}{Numeric vector for the x-axis}

\item{y}{Numeric vector for the y-axis}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{main}{Main plot title, overrides default correlation}

\item{main.prefix}{Text to add to plot title before correlation}

\item{size}{Point size}

\item{col}{Color for points}

\item{fixed}{Should axes be fixed to the same x and y}

\item{alpha}{Point transparency}

\item{shape}{Point shape (integer code passed to ggplot2)}

\item{lm.col}{Color of the line}

\item{do.lm}{Should a linear fit be rendered}

\item{method}{Method for fitting the line. Anything supported by \code{\link{ggplot2::geom_smooth()}}}

\item{lm.group}{Group to fit model in}

\item{raster}{Should points be rasterized (usefull for large numbers of points)}

\item{dpi}{When raster=TRUE the DPI to render at}

\item{facet}{Vector with facet variable}

\item{facet.ncol}{Number of columns in facet}
}
\value{
A ggplot2 object
}
\description{
Scatterplot using ggplot
}
