% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{read.cellprofiler.fileset.b}
\alias{read.cellprofiler.fileset.b}
\title{Read a cell level fileset type B}
\usage{
read.cellprofiler.fileset.b(
  prefix,
  return.feature.meta = F,
  add.global.id = T,
  merging.strategy = "mean",
  parent.col = "Parent_cell",
  pat.exp = ".*Experiment.txt",
  pat.img = ".*Image.txt",
  pat.cells = ".*cell.txt",
  pat.orl = ".*Object relationships.txt",
  pat.others = "^.*_([a-z]+\\\\d*).txt$",
  na.rm = F
)
}
\value{
list with data frames:
- cells (cell level features)
- meta (image level features)
- objectRelations 
- features [optional]
Output is NULL if no cells are detected.
}
\description{
Reads a zip file with cellprofiler features, each file other then
_Image, _Experiment and _Object Relationships are assumed to be an object.
Will match objects on order with an appropriate matching strategy.
}
