% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{scale_assay_min_max}
\alias{scale_assay_min_max}
\title{Scale a TglowAssay between a min and max value for each feature}
\usage{
scale_assay_min_max(
  assay,
  slot = "data",
  min = 0,
  max = 1,
  scale.by.var = FALSE,
  scaling.factors = NULL
)
}
\arguments{
\item{assay}{A \linkS4class{TglowAssay}}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{min}{The minimal value for each feature in output}

\item{max}{The maximum value for each feature in output}

\item{scale.by.var}{Multiply the result by the orignal variance}

\item{scaling.factors}{Vector of scaling factors per feature}
}
\value{
A \linkS4class{TglowAssay} with the data slot populated with the rescaled data.
}
\description{
Scale the data slot between min and max and return a new assay
}
