% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{scale_grouped}
\alias{scale_grouped}
\title{Z-score or modified z-score a vector per group}
\usage{
scale_grouped(data, grouping, method = "mod.z")
}
\arguments{
\item{data}{A numeric vector}

\item{grouping}{A vector of the same length as data, the groups to scale within}

\item{method}{Either "mod.z" for a modified z-score (see \code{\link{mod_zscore}}) or "z" for a z-score}
}
\value{
The vector scaled within each group
}
\description{
Z-score or modified z-score a vector per group
}
