% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.r
\name{tglow.dataset.from.list}
\alias{tglow.dataset.from.list}
\title{Tglow dataset from list}
\usage{
tglow.dataset.from.list(
  output,
  assay,
  meta.patterns = c("ImageNumber", "ObjectNumber", "Object_Number", "Parent"),
  img.feature.patterns = c("^AreaOccupied", "^ImageQuality", "^Granularity"),
  col.object = "cell_ObjectNumber_Global",
  col.img.id = "cell_ImageNumber_Global",
  col.meta.img.id = "ImageNumber_Global"
)
}
\arguments{
\item{output}{Tglow list object obtained from tglow.read.dir}

\item{assay}{Which assay to convert}

\item{meta.patterns}{Grep patterns in collumn names asspciated with object level metadata items}

\item{img.feature.patterns}{Grep patterns in image column names assocated with image level features to analyze}

\item{col.object}{The collumn name in the features which contains the per object object identifier}

\item{col.img.id}{The collumn name in the features which contains the per object image identifier}

\item{col.meta.img.id}{The collumn name in the image level data which contains the image id}
}
\value{
A populated TglowDataset
}
\description{
Tglow dataset from list
}
