% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.r
\name{tglow_assay_from_list}
\alias{tglow_assay_from_list}
\title{Convert a tglow list to a tglow assay}
\usage{
tglow_assay_from_list(output, assay, meta.cols, col.object)
}
\arguments{
\item{output}{Tglow list object obtained from tglow.read.dir}

\item{assay}{Which assay to convert}

\item{meta.cols}{Which object level collumns in the features to consider as metadata}

\item{col.object}{Which collumn to use as the identifier. Must be in meta.cols}
}
\description{
Convert a tglow list to a tglow assay
}
