% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_image.r
\name{tglow_build_img_index}
\alias{tglow_build_img_index}
\title{Build an index of where the source images are stored}
\usage{
tglow_build_img_index(path, plate_filter = NULL, pattern = ".ome.tiff$")
}
\arguments{
\item{path}{path to tglow image dir}

\item{plate_filter}{plate names to run}

\item{pattern}{the pattern to match image names and then extract the fields (gsubbed away)}
}
\value{
A data frame with the image paths
}
\description{
This assumes images are organized as follows:
\if{html}{\out{<plate>}}/\if{html}{\out{<row>}}/\if{html}{\out{<col>}}/\if{html}{\out{<field>}}.ome.tiff

Matches images based on \if{html}{\out{<plate>}}/\if{html}{\out{<row>}}/\if{html}{\out{<col>}}/\if{html}{\out{<field>}}.ome.tiff
}
