% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{tglow_dimplot}
\alias{tglow_dimplot}
\title{Plot a reduction}
\usage{
tglow_dimplot(
  object,
  reduction,
  ident = NULL,
  assay = NULL,
  slot = NULL,
  downsample = NULL,
  log.ident = FALSE,
  axis.x = 1,
  axis.y = 2,
  xlab = NULL,
  ylab = NULL,
  no.colscale = FALSE,
  labs = NULL,
  labs.add = TRUE,
  labs.size = 5,
  labs.textcol = "white",
  labs.bgcol = "black",
  ...
)
}
\arguments{
\item{object}{A \linkS4class{TglowDataset}}

\item{reduction}{A name of a reduction on dataset.}

\item{ident}{The item to use for coloring points, passed to \code{\link{getDataByObject}}}

\item{assay}{The assay to use for coloring, passed to \code{\link{getDataByObject}}}

\item{slot}{The slot to use for coloring, passed to \code{\link{getDataByObject}} Can be "data" or "scale.data"}

\item{downsample}{Should downsampling be applied prior to plot. NA's are removed first}

\item{log.ident}{Should color vector be log2 transformed}

\item{axis.x}{Which dimension from reduction to plot in x}

\item{axis.y}{Which dimension from reduction to plot in y}

\item{xlab}{Override xlab}

\item{ylab}{Override ylab}

\item{no.colscale}{Skip adding colorscale}

\item{labs}{Matrix with x, y and label for ident labels to add, autogenerated for factors and character idents under 50 classes}

\item{labs.add}{Should labels be added to the plot}

\item{labs.size}{Point size of text labels}

\item{labs.textcol}{Text color of the labels}

\item{labs.bgcol}{Color of the background behind the text}

\item{...}{Remaining arguments passed to \code{\link{plot_xy}}}
}
\description{
Plot a reduction on a \linkS4class{TglowDataset}
}
