% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{tglow_example}
\alias{tglow_example}
\title{Imaging features on drug perturbed T cells}
\format{
A \linkS4class{TglowDataset} with 5000 cells and 989 features
\itemize{
  \item assays: Tglow assays - just has a raw assay
}
}
\description{
A \linkS4class{TglowDataset} dataset with data on drug perturbed T cells.
}
