% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_image.r
\name{tglow_read_binmat}
\alias{tglow_read_binmat}
\title{Read a binary matrix, used for reading the registration matrices}
\usage{
tglow_read_binmat(path)
}
\arguments{
\item{path}{Path to the binary matrix file}
}
\description{
Adapted from: https://www.r-bloggers.com/2012/06/getting-numpy-data-into-r/
}
