% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{TglowAssay-class}
\alias{TglowAssay-class}
\title{TglowAssay}
\description{
TglowAssay
}
\section{Slots}{

\describe{
\item{\code{data}}{Slot to store unscaled HCI features. Rows are objects, columns are features}

\item{\code{scale.data}}{Slot to store a scaled version of data}

\item{\code{features}}{Data frame with feature level metadata}
}}

