% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.r
\name{TglowAssayFromMatrix}
\alias{TglowAssayFromMatrix}
\title{Create a new TglowDataset}
\usage{
TglowAssayFromMatrix(objects, scaled.objects = NULL, features = NULL)
}
\arguments{
\item{objects}{The object level matrix}

\item{scaled.objects}{The scaled object level matrix (optional)}

\item{features}{The data frame with features}
}
\value{
A TglowAssay
}
\description{
Create a new TglowDataset
}
