% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{TglowDataset-class}
\alias{TglowDataset-class}
\title{TglowDataset}
\description{
Object to store HCI features along side object (cell) and image level metadata.
}
\section{Slots}{

\describe{
\item{\code{assays}}{List of TglowAssays containing the data}

\item{\code{meta}}{Data frame containing object (cell) level metadata}

\item{\code{object.ids}}{Vector containing object (cell) level identifiers}

\item{\code{image.meta}}{Image level metadata}

\item{\code{image.data}}{TglowAssay object with image level features}

\item{\code{image.ids}}{Vector of length objects with corresponding image ids}

\item{\code{reduction}}{List to store PCA/UMAP embeddings}

\item{\code{graph}}{Slot to store kNN/AkNN graph}

\item{\code{active.assay}}{The default assay to analyze. Not all functions may respect this}
}}

