% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.r
\name{TglowDatasetFromList}
\alias{TglowDatasetFromList}
\title{Tglow dataset from list}
\usage{
TglowDatasetFromList(
  output,
  assay,
  meta.patterns = c("ImageNumber", "ObjectNumber", "Object_Number", "Parent",
    "_Location_", "BoundingBox"),
  img.feature.patterns = c("^AreaOccupied", "^ImageQuality", "^Granularity"),
  col.object = "cell_ObjectNumber_Global",
  col.img.id = "cell_ImageNumber_Global",
  col.meta.img.id = "ImageNumber_Global"
)
}
\arguments{
\item{output}{Tglow list object obtained from \code{\link[=read_cellprofiler_dir]{read_cellprofiler_dir()}}}

\item{assay}{Which assay to convert (name of the list item)}

\item{meta.patterns}{Grep patterns in column names associated with object level metadata items}

\item{img.feature.patterns}{Grep patterns in image column names associated with image level features to analyze}

\item{col.object}{The column name in the features which contains the per object object identifier}

\item{col.img.id}{The column name in the features which contains the per object image identifier}

\item{col.meta.img.id}{The column name in the image level data which contains the image id}
}
\value{
A populated TglowDataset
}
\description{
Tglow dataset from list
}
