% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.r
\name{TglowDatasetFromMatrices}
\alias{TglowDatasetFromMatrices}
\title{Create a new TglowDataset}
\usage{
TglowDatasetFromMatrices(
  objects,
  images,
  image.ids,
  object.meta = NULL,
  image.meta = NULL,
  assay.out = "raw"
)
}
\arguments{
\item{objects}{The object level data matrix, must be a matrix, and be numeric}

\item{image.ids}{The of nrow(objects) de which rownames of image corresponds to which object}

\item{object.meta}{A data.frame with metadata, one row per object. Defaults to data.frame(id=rownames(objects))}

\item{image.meta}{A data.frame with metadata, one row per image. Defaults to data.frame(id=rownames(images))}

\item{assay.out}{The assay name to store objects under}

\item{image}{The image level data matrix, must be a matrix, and be numeric}
}
\value{
A populated TglowDataset
}
\description{
Create a new TglowDataset object from a matrix with image level data and a matrix of object level data
}
\details{
Only the data slot on assay is populated with the values in \code{objects} if this is not what you want you can
always re-assign data to the scale.data slot afterwards by \code{dataset$raw@scale.data <- dataset$raw@data}.
Just to note, the assumption is made that the scale.data slot is actually scaled to mean=0 variance=1 when
running PCA's etc, so overriding with non-scaled data may change your results.
}
