% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.r
\name{TglowFeatureLocation}
\alias{TglowFeatureLocation}
\title{Create a new empty TglowFeatureLocation}
\usage{
TglowFeatureLocation(feature, assay = NULL, slot = NULL)
}
\arguments{
\item{feature}{Character with the feature position}

\item{assay}{Assay to grab feature from, or NULL if metadata feature}

\item{slot}{slot to grab feature from, or NULL if metadata feature}
}
\description{
Create a new empty TglowFeatureLocation
}
