% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{TglowFeatureMap-class}
\alias{TglowFeatureMap-class}
\title{TglowFeatureMap}
\description{
TglowFeatureMap
}
\section{Slots}{

\describe{
\item{\code{feature.x}}{Feature describing object X}

\item{\code{feature.y}}{Feature describing object Y}

\item{\code{feature.z}}{Feature describing object Z}

\item{\code{feature.well}}{Feature describing well position}

\item{\code{feature.plate}}{Feature describing plate}

\item{\code{feature.field}}{Feature describing field}
}}

