% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{TglowFilter-class}
\alias{TglowFilter-class}
\title{TglowFilter}
\description{
An object to describe a filter for either objects (cells) or
features
}
\section{Slots}{

\describe{
\item{\code{name}}{Filter name}

\item{\code{column_pattern}}{Collumn pattern to apply filters on. "all" represents all}

\item{\code{func}}{Name of filter function}

\item{\code{threshold}}{Filter threshold value}

\item{\code{active}}{Logical indicating if the filter should be applied}

\item{\code{transpose}}{Should data matrix be transposed prior to running func}
}}

