% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{TglowReduction-class}
\alias{TglowReduction-class}
\title{TglowReduction}
\description{
TglowReduction
}
\section{Slots}{

\describe{
\item{\code{x}}{Slot to store PCA / UMAP coordinates}

\item{\code{var}}{optional variances (squared standard deviations)}

\item{\code{var_total}}{optional total variance (sum of var)}

\item{\code{object}}{Optional PCA / UMAP object}
}}

