% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{add_global_ids}
\alias{add_global_ids}
\title{Add a global id to a matrix of image files}
\usage{
add_global_ids(matrix, fileset.id)
}
\arguments{
\item{matrix}{An input matrix or data.frame}

\item{fileset.id}{The global fileset id to add}
}
\value{
A data.frame with extra columns suffixed by _Global with globally unique ids
}
\description{
Take a matrix and extract patterns 'ImageNumber', 'ObjectNumber' and
'Object_Number' and add a globably unique prefix. Will store these globally unique
id's in columns \if{html}{\out{<original>}}_Global.
}
\details{
Matrix with column pattern 'ImageNumber', 'ObjectNumber' and 'Object_Number'
which will be duplicated and have a globally unique variable added
Output of this is stored in the same column name but with suffix _Global
}
