% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{apply_boxcox}
\alias{apply_boxcox}
\title{Transform assay data using boxcox transform}
\usage{
apply_boxcox(
  dataset,
  assay,
  assay.out = NULL,
  trim = TRUE,
  slot = "data",
  verbose = TRUE,
  rfast.zerotol = 1e-10,
  ...
)
}
\arguments{
\item{dataset}{A tglow dataset}

\item{assay}{The assay to use.}

\item{assay.out}{The name of the output assay. Defaults to paste0(assay, "-trans")}

\item{trim}{Logical indicating to trim zero variance columns and NA's after applying boxcox transform}

\item{slot}{The slot to use for calculating filters, defaults to "data". Can be "data" or "scale.data"}

\item{verbose}{Raise a warning if columns are trimmed}

\item{rfast.zerotol}{threshold to consider variances zero in Rfast see details.}

\item{...}{Arguments passed to \code{\link{boxcox_transform}}}
}
\value{
Tglow object with new boxcox assay
}
\description{
Transform a assay using boxcox transform. If assay="image.data", result is stored in slot image.data.trans.
Lambda values are added to the features data.frame: 0=log, 1=no transformation
}
\details{
NOTE: Currently NOT using Rfast::colVars() as there are inconsistencies with matrixStats::colVars()
Rfast's does not return zero variances, so we have to set some minial value to consider as 0
https://github.com/RfastOfficial/Rfast/issues/62
https://github.com/RfastOfficial/Rfast/issues/115
}
