% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.r
\name{apply_feature_filters}
\alias{apply_feature_filters}
\title{Apply feature filters to a tglow dataset}
\usage{
apply_feature_filters(dataset, filter.res, assays = NULL)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{filter.res}{The matrix of filter output from \code{\link{calculate_feature_filters}} or a named logical vector}

\item{assays}{Only apply filteres to these assays}
}
\value{
The filtered \linkS4class{TglowDataset}
}
\description{
Apply a logical matrix or a logical array with filters on the columns and features on the rows
to a tglow dataset. Each assay is seperately subsetted to remove features
which are not selected.

If filter.res is a matrix, all rows must be true.
}
