% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{boxcox_transform}
\alias{boxcox_transform}
\title{Transform assay data using boxcox transform}
\usage{
boxcox_transform(
  x,
  return.lambda = FALSE,
  limit = 5,
  fudge = 0.1,
  downsample = NULL,
  add.lambda = FALSE
)
}
\arguments{
\item{x}{A numeric vector with values to transform}

\item{return.lambda}{return the lambda value instead}

\item{limit}{The absolute value of lambda's to test}

\item{fudge}{Fudgefactor to apply either log transform, or just leave it as is}

\item{downsample}{Calculate labmda's on a random sample of the data. Either NULL (no downsampling), an integer (selects x values), or a selection vector}

\item{add.lambda}{Return a list with two items instead, one transformed x, two, the lamda used}
}
\value{
A numeric vector with the boxcox transformed data or the optimal lambda value
}
\description{
Transform a vector using boxcox transform
If value is not positive it is offset to make everything positive
}
