% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.r
\name{calculate_eigenfeatures}
\alias{calculate_eigenfeatures}
\title{Calculate the eigenfeature (first PC) for a cluster of features}
\usage{
calculate_eigenfeatures(dataset, assay, slot, cluster.col)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay to use for clustering}

\item{slot}{The data slot to use ("data", "scale.data", etc.)}

\item{cluster.col}{Column in features slot containing cluster assignments to calculate eigenfeatures for}
}
\value{
Returns the \linkS4class{TglowDataset} with a new assay named "\if{html}{\out{<assay>}}_eigenfeatures"
containing the eigenfeature for each cluster. The features slot contains the variance explained by the eigenfeature.
}
\description{
Calculates eigenfeatures for a set of features defined by a column (see calculate_feature_clustering).
The eigenfeature is the first principal component of the features in the cluster.
This assay can be used for more fair testing of how well a featureset is associated to a trait.
}
