% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{calculate_lm}
\alias{calculate_lm}
\title{Calculate linear coefficients}
\usage{
calculate_lm(
  dataset,
  assay,
  slot,
  covariates,
  formula = NULL,
  grouping = NULL,
  assay.covar = NULL,
  slot.covar = NULL,
  assay.image = NULL,
  covariates.dont.use = NULL,
  rescale.group = FALSE,
  ...
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay on dataset to use}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{covariates}{Character vector of independent variables to use in the model}

\item{formula}{The formula to use for regression, string or formula. Defaults to additive model. See details}

\item{grouping}{Vector with grouping variable if residuals be calculated per group of objects. See details}

\item{assay.covar}{The assay to grab covariates from. Defaults to assay argument}

\item{slot.covar}{The slot to grab covariates from. Can be "data" or "scale.data"}

\item{assay.image}{The image assay to use for grabbing covariates, NULL, "image.data", "image.data.trans" or "image.data.norm"}

\item{covariates.dont.use}{Only use if you understand the implications. See details}

\item{rescale.group}{When grouping is active, should the group be re-centered and scaled prior to regressing}

\item{...}{Remaining arguments passed to \code{\link[=lm_matrix]{lm_matrix()}}. Use with caution.}
}
\value{
A list of regression results. If grouping != NULL, there is one list per group
}
\description{
Fit a linear model using OLS and find coefficients
}
\details{
\code{grouping}

Runs seperate regressions per group in the indicated grouping variable. length(grouping) must be nrow(assay)

\code{formula}

If NULL an additive model of all covariates is performed. Otherwise should be a string interpretable by \code{\link[stats:formula]{stats::formula()}}

\code{covariates.dont.use}

Fit the model with these covariates included, but don't use them for calculating residuals (affects r2, se etc)
The covariates for these variables are removed in the output. When specifying more complex models in formula
use the term names, with for instance, interaction terms for example, if it has a form of '~ a + b + c + b:c' and you don't
want to consider the interaction term, add 'b:c'. To remove the intercept, add '(Intercept)'
}
