% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{calculate_lmm}
\alias{calculate_lmm}
\title{Calculate linear coefficients using a linear mixed model}
\usage{
calculate_lmm(
  dataset,
  assay,
  slot,
  covariates,
  formula = NULL,
  grouping = NULL,
  assay.covar = NULL,
  slot.covar = NULL,
  assay.image = NULL,
  rescale.group = TRUE,
  ...
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay on dataset to use}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{covariates}{Character vector of independent variables to use in the model}

\item{formula}{The formula to use for regression. Defaults to additive model. See details}

\item{grouping}{Vector with grouping variable if residuals be calculated per group of objects. See details}

\item{assay.covar}{The assay to grab covariates from. Defaults to assay argument}

\item{slot.covar}{The slot to grab covariates from. Can be "data" or "scale.data"}

\item{assay.image}{The image assay to use for grabbing covariates, NULL, "image.data", "image.data.trans" or "image.data.norm"}

\item{rescale.group}{When grouping is active, should the subset be re-centered and scaled prior to regressing}

\item{...}{Remaining arguments passed to \code{\link[=lmm_matrix]{lmm_matrix()}} and then to \code{\link[lmerTest:lmer]{lmerTest::lmer()}}. See details}
}
\value{
A list of regression results. If grouping != NULL, there is one list per group
}
\description{
Fit a linear mixed model using REML / MLE (lme4) and find coefficients
}
\details{
Strongly recommend reading extra parameters available in \code{\link[=lmm_matrix]{lmm_matrix()}}

\code{grouping}

Runs seperate regressions per group in the indicated grouping variable. length(grouping) must be nrow(assay)

\code{formula}

Should be a string interpretable by \code{\link[stats:formula]{stats::formula()}} and \code{\link[lmerTest:lmer]{lmerTest::lmer()}} but should NOT
have response in the forumla. E.g. when regressing \code{y ~ x + (1|donor)} forumula should be \code{~ x + (1|donor)}
Defaults to a regular additive linear model.
To specify derived terms you might need to add I(), for polynomials for example use either poly() or I(x^2)
}
