% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.r
\name{calculate_object_filters}
\alias{calculate_object_filters}
\title{Apply a cell level filter to a tglow dataset}
\usage{
calculate_object_filters(
  dataset,
  filters,
  assay,
  slot = "data",
  grouping = NULL,
  features = NULL,
  na.fail = TRUE
)
}
\arguments{
\item{dataset}{A tglow dataset}

\item{filters}{A list with TglowFilter objects to apply}

\item{assay}{The assay to use for calculating filters}

\item{slot}{The slot to use for calculating filters: "data" or "scale.data"}

\item{grouping}{A vector specifying the grouping for some filters}

\item{features}{An optional subset of features to use for calculation.}

\item{na.fail}{Should NA be treated as fail, defaults to yes}
}
\value{
A logical matrix of nrow(dataset) x length(filters) where T indicates filter pass and F indicates filter fail.
}
\description{
Calculate an object filter table on a \linkS4class{TglowDataset}
All filters are inclusive
}
