% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.r
\name{calculate_pca}
\alias{calculate_pca}
\title{Calculate the PCA on an assay}
\usage{
calculate_pca(
  dataset,
  assay,
  slot = "scale.data",
  pc.n = NULL,
  reduction.name = NULL,
  ret.prcomp = FALSE,
  use_irlba = FALSE,
  recenter = FALSE,
  rescale = FALSE
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay to use}

\item{slot}{The slot to use for calculating filters, defaults to "data". Can be "data" or "scale.data"}

\item{pc.n}{How many PCs to calculate (if not NULL uses \code{\link[=prcomp_irlba]{irlba::prcomp_irlba()}} instead of \code{\link[=prcomp]{prcomp()}}}

\item{reduction.name}{The name to save the PCA results under. Defualts to PCA-\if{html}{\out{<assay>}}}

\item{ret.prcomp}{Instead of returning just the PCs and variances, return the whole prcomp object}

\item{use_irlba}{Logical if \code{\link[=prcomp_irlba]{irlba::prcomp_irlba()}} or \code{\link[=prcomp]{prcomp()}} should be used for PCA}

\item{recenter}{Logical if matrix will be recentered. This is advisable if using modified z-score}

\item{rescale}{Logical if matrix will be rescaled. This is advisable if using modified z-score}
}
\value{
\linkS4class{TglowDataset} with populated PCA slot
}
\description{
Calculate a principal component analysis on a \linkS4class{TglowDataset} and add it to
the reductions. Does not re-scale by default and assumes scale.data
is scaled has mean 0 variance 1, which is not true if using modified zscore
}
