% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{calculate_sumstats}
\alias{calculate_sumstats}
\title{Calculate summary stats on features}
\usage{
calculate_sumstats(dataset, assay, slot)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay on dataset to use}

\item{slot}{The assay slot to use ("data", "scale.data")}
}
\value{
The \linkS4class{TglowDataset}
}
\description{
Calculate the mean, median, mode, skewness, kurtosis, geometric mean, min, max
Removes NA's for each feature individually.
Stats are put on the assay@features slot under the name 'slot'_'stat'
}
