% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.r
\name{calculate_umap}
\alias{calculate_umap}
\title{Calculate the UMAP on an assay}
\usage{
calculate_umap(
  dataset,
  reduction,
  assay = NULL,
  slot = "scale.data",
  pc.n = 30,
  reduction.name = NULL,
  downsample = NULL,
  use_irlba = TRUE,
  ...
)
}
\arguments{
\item{dataset}{A tglow dataset}

\item{reduction}{The reduction to use for calculating UMAPs. If NULL re-calculated}

\item{assay}{The assay to use}

\item{slot}{The slot to use for calculating filters, defaults to "data". Can be "data" or "scale.data"}

\item{pc.n}{How many PC's to calculate}

\item{reduction.name}{The name to save the UMAP results under. Defualts to UMAP-<assay>}

\item{downsample}{Downsample to a random subset of objects prior to running UMAP. Can be an integer or a selection vector of row ids.}

\item{use_irlba}{Logical if \code{\link[=prcomp_irlba]{irlba::prcomp_irlba()}} or \code{\link[=prcomp]{base::prcomp()}} should be used for PCA}

\item{...}{Arguments passed to \code{\link{uwot::umap()}}}
}
\value{
\linkS4class{TglowDataset} with populated reduction slot
}
\description{
Calculate UMAP on a tglow dataset and add it to the reductions.
}
\details{
Uses an existing reduction if possible. When downsampling objects it returns
a matrix in the reductions slot which has nrow(dataset) with objects which were
omitted set to NA to ensure downstream ordering is maintained.
}
