% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{correct_lm_per_featuregroup}
\alias{correct_lm_per_featuregroup}
\title{Linearly correct for a set of covariates}
\usage{
correct_lm_per_featuregroup(
  object,
  assay,
  slot,
  covariates.group,
  slot.covar = NULL,
  assay.covar = NULL,
  assay.image = NULL,
  assay.out = NULL,
  grouping = NULL,
  covariates.dont.use = NULL,
  rescale.group = FALSE,
  na.rm = F
)
}
\arguments{
\item{object}{A \linkS4class{TglowDataset}}

\item{assay}{The assay to use}

\item{slot}{The slot to use for regressing against. Can be "data" or "scale.data"}

\item{covariates.group}{List with specific covariates for groups of features. See detaills}

\item{slot.covar}{The slot to grab covariates from. Can be "data" or "scale.data". Defaults to slot}

\item{assay.covar}{The assay to grab covariates from. Defaults to assay argument}

\item{assay.image}{The image assay to use for grabbing covariates, NULL, "image.data", "image.data.trans" or "image.data.norm"}

\item{assay.out}{Name of the output assay. Defaults to \if{html}{\out{<assay>}}.lm.corrected}

\item{grouping}{Vector with grouping variable if residuals be calculated per group of objects. See details}

\item{covariates.dont.use}{Vector of covariate names to NOT use when calculating residuals. See detaills}

\item{rescale.group}{When grouping is active, should the subset be re-centered and scaled prior to regressing}

\item{na.rm}{Remove NA's from the design matrix prior to regressing (does not remove NAs in the response)}
}
\value{
The \linkS4class{TglowDataset} with a corrected assay
}
\description{
Fit a linear model using OLS and correct an assay for specified covariates
}
\details{
\code{covariates.group}

This must be a list where the name of the list item is a grep pattern of the features to apply the covariates to, and the item is a vector of feature names to correct for

\code{grouping}

If this is provided, scaling for populating scale.data slot is done over ALL residuals, not per group to ensure the mean and sd of the whole vector is as expected

\code{covariates.dont.correct}

The beta's for these variables are removed when calculating the residuals. When specifying more complex models in formula,
use the term names, with for instance, interaction terms for example, if it has a form of '~ a + b + c + b:c' and you don't
want to consider the interaction term, add 'b:c'. To remove the intercept, add '(Intercept)'
}
