% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_image.r
\name{depr_tglow_read_imgs_aics}
\alias{depr_tglow_read_imgs_aics}
\title{Load images around the center of a cell object using AICSImageIO}
\usage{
depr_tglow_read_imgs_aics(
  dataset,
  cell.subset,
  python,
  group.col,
  img.index,
  reg.index = NULL,
  window = 75,
  feature.x = "cell_AreaShape_Center_X",
  feature.y = "cell_AreaShape_Center_Y",
  channels = NULL,
  planes = NULL,
  max.project = T,
  assay = "raw",
  slot = "data"
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{cell.subset}{The subset of cells to retrieve}

\item{python}{python install with AICSImageIO}

\item{group.col}{The feature to match cells to img.index. Usually 'Metadata_group'}

\item{img.index}{A matrix with image paths generated by tglow.build.img.index}

\item{reg.index}{A matrix with registration paths generated by tglow.build.img.index
Must have an additional column named 'plate2' to indicate the registration plate}

\item{window}{Window in px around the cell center to retrieve}

\item{feature.x}{The feature that describes the object pos in px in x}

\item{feature.y}{The feature that describes the object pos in px in y}

\item{channels}{The channels to read. A list of Vector of indices per cycle. (default NULL = all channels)}

\item{planes}{The planes to read. A list of Vector of indices per cycle. (default NULL = all planes)}

\item{max.project}{Should the stack be max projected per channel? (default TRUE)}

\item{assay}{The assay on dataset to use}

\item{slot}{The assay slot to use ("data", "scale.data")}
}
\value{
A list of EBImage objects by object id
}
\description{
Same as tglow_read_imgs but uses AICSImageIO as the backend. This is slower then tglow_read_imgs.
}
