% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{effective_dimensionality}
\alias{effective_dimensionality}
\title{Caclulate the effective dimensionality of a dataset}
\usage{
effective_dimensionality(data, method = "LiJi", var.thresh = 0.95)
}
\arguments{
\item{data}{A matrix with data}

\item{method}{'LiJi', 'pc_var', 'Chevrud', 'Galwey' . See details}

\item{var.test}{Percentage of variance cutoff}
}
\description{
Calculates the effective dimensionality of a multidimensional dataset.
}
\details{
\code{method}

\code{eigenval} are the eigenvalues of the correlation matrix.

\code{m} are the number of eigenvalues
\itemize{
\item \code{LiJi} Use Li and Ji's method - https://www.nature.com/articles/jhg201134#Sec2

\code{eff.tests <- sum(I( eigenval > 1) + ( eigenval - floor( eigenval)))}
\item \code{pc_var} Number of PC's needed to pass var.thresh

\code{eff.tests <- which(cumsum(eigenval / sum(eigenval)) >= var.thresh)}
\item \code{Chevrud}:

\code{eff.tests <- m +(m - 1) * (1-(var(eigenval) / m))}
\item \code{Galwey}
}
}
