% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{effective_dimensionality}
\alias{effective_dimensionality}
\title{Calculate the effective dimensionality of a dataset}
\usage{
effective_dimensionality(data, method = "li_ji", var.thresh = 0.95)
}
\arguments{
\item{data}{A matrix with data}

\item{method}{'all', 'li_ji', 'pc_var', 'cheverud', 'galwey' . See details}

\item{var.test}{Percentage of variance cutoff}
}
\value{
A list with the results if method='all' or the value of effective dimesionality.
}
\description{
Calculates the effective dimensionality of a multidimensional dataset.
}
\details{
\code{method}

\code{eigenval} are the eigenvalues of the correlation matrix.

\code{m} are the number of eigenvalues

Formulas derived from: https://www.nature.com/articles/jhg201134#Sec2
\itemize{
\item \code{li_ji} Use Li and Ji's method

\code{eff.tests <- sum(I( eigenval > 1) + ( eigenval - floor( eigenval)))}
\item \code{pc_var} Number of PC's needed to pass var.thresh

\code{eff.tests <- which(cumsum(eigenval / sum(eigenval)) >= var.thresh)}
\item \code{cheverud}:

\code{eff.tests <- 1 +(m - 1) * (1-(var(eigenval) / m))}
\item \code{galwey}:

eff.tests <- sum(sqrt(eigenval))^2 / sum(eigenval)
}
}
