% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_rep_objects.r
\name{fetch_representative_object_nn}
\alias{fetch_representative_object_nn}
\title{Retrieve a cell (and its neighbours) based on a feature sumstat}
\usage{
fetch_representative_object_nn(
  dataset,
  feature,
  reduction,
  assay = NULL,
  slot = NULL,
  n = 0
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{feature}{The feature to find a representative objects for, used as a grouping variable}

\item{reduction}{The reduction to use to calculate distance. Reccomend a PCA.}

\item{assay}{if feature is on an assay, the assay to use}

\item{slot}{if feature is on an assay, which slot to use}

\item{n}{How many nearest neighbours to the central most object should be fetched}
}
\value{
vector of indices in objects matrix
}
\description{
Gets a objects and its closes neighbours based on a reduction and a feature.
Feature should be catagorical, can be used to find the object that is closest to the centroid
of a group in reduction space. To get more than one object per group, specify n.
}
