% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{fetch_representative_object_quantiles}
\alias{fetch_representative_object_quantiles}
\title{Retrieve a set of objects (and its neighbours) based on a feature sumstat}
\usage{
fetch_representative_object_quantiles(
  dataset,
  assay,
  slot,
  feature,
  name = NULL,
  n = 1
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay to use}

\item{slot}{The slot to use for calculating filters, defaults to "data". Can be "data" or "scale.data"}

\item{feature}{The feature to find a representative cell for}

\item{name}{Prefix to add to the names of the names vector}

\item{n}{How many objects either side of the representitative cell should be returend}
}
\value{
A list with row ids and labels for the objects
}
\description{
Fetches (n*2)+1 objects arround the 0th, 10th, 25th, 50th, 75th, 90th and 100th quantiles
}
