% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{find_markers_lmm}
\alias{find_markers_lmm}
\title{Find marker features using a linear (mixed) model}
\usage{
find_markers_lmm(
  dataset,
  ident,
  assay,
  slot,
  covariates = NULL,
  assay.covar = NULL,
  slot.covar = NULL,
  formula = NULL,
  contrast = NULL,
  ref.classes = NULL,
  assay.image = NULL,
  return.top = NULL,
  method = "lmm",
  control = lme4::lmerControl(check.conv.singular = list(action = "ignore", tol = 1e-04))
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{ident}{A column in meta, image.meta, assay, assay.image to use as class labels}

\item{assay}{The assay on dataset to use}

\item{slot}{The assay slot to use ("data", "scale.data")}

\item{covariates}{Character vector of covariates to use in model}

\item{assay.covar}{The assay to grab covariates from}

\item{slot.covar}{The slot to grab covariates from}

\item{formula}{Formula for the mixed model}

\item{contrast}{Contrast matrix for comparisons, see details}

\item{ref.classes}{Reference classes to compare against}

\item{assay.image}{The image assay to use for grabbing ident}

\item{return.top}{Return the top x values by abs(t-stat). Defaults to NULL (all)}

\item{method}{Method to use, either "lm" or "lmm"}

\item{control}{Control parameters passed to lmerControl()}
}
\value{
A data frame with mixed model results
}
\description{
Find marker features using a linear (mixed) model
}
\details{
Defining the contrast matrix:
The contrast matrix should have dimensions n x m where n is the number of levels in ident to test
and m is the number of parameter in the model, includding any fixed effect covariates.
Each row is treated as a separate test and must sum to 1. Positive class should be 1, reference
classes should sum to -1 and all other classes should be 0.

If NULL (default) a one-vs-all contrast matrix is created.
If NULL and ref.classes is provided, the contrast is one-vs-ref.classes.

Setting lmerControl:
By default we set check.conv.singular to ignore, as singular fits are common for HCI features with low variance
and these warings can mess up the progress bar.
Rather then a warning, the output matrix will contain a column 'singular_reff' indicating if the fit was singular.
}
