% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{find_markers_ttest}
\alias{find_markers_ttest}
\title{Find marker features using a t-test}
\usage{
find_markers_ttest(
  dataset,
  ident,
  assay,
  slot,
  assay.image = NULL,
  return.top = NULL,
  ref.classes = NULL,
  na.rm = T
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{ident}{A column in meta, image.meta, assay, assay.image to use as class labels}

\item{assay}{The assay to use}

\item{slot}{The slot to use for finding markers. Can be "data" or "scale.data"}

\item{assay.image}{The image assay to use for grabbing ident, NULL, "image.data", "image.data.trans" or "image.data.norm"}

\item{return.top}{Return the top x values by pvalue. Defaults to 10, set to NULL to return all}

\item{ref.classes}{A vector of values to compare class mean against. Default NULL uses all other classes as reference}

\item{na.rm}{Should NA's be removed during t-test}
}
\value{
A data.frame with t-test results
}
\description{
Find marker features using a two sample t-test. By default top 10 positive results
for a class are returned
}
