% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.r
\name{find_outliers_pca}
\alias{find_outliers_pca}
\title{Find outliers in PCA space}
\usage{
find_outliers_pca(
  dataset,
  assay,
  qc.group = NULL,
  thresh = 3.5,
  pc.thresh = 0.75,
  pc.max = NULL,
  pc.n = NULL,
  slot = "data",
  method = "z",
  return.pcs = FALSE,
  use_irlba = TRUE
)
}
\arguments{
\item{dataset}{A tglow dataset}

\item{assay}{The assay to use}

\item{qc.group}{A vector indicating if PC's should be calculated in subgroups of the data. Default find outliers using all objects at once}

\item{thresh}{Threshold in absolute PC to consider an outlier}

\item{pc.thresh}{The percentage of variance of PC's to select for outlier detection}

\item{pc.max}{The maximum number of components to calculate using \code{\link[=prcomp_irlba]{irlba::prcomp_irlba()}}}

\item{pc.n}{The number of PC's to use. Defaults to the number of PC's that reach pc.thresh or pc.max}

\item{method}{Method to scale PC's prior to selecting thresh. Value can be 'z' for z-score or 'mod.z' for modified zscore}

\item{return.pcs}{Should the grouped PC's be returned?}

\item{use_irlba}{Logical if \code{\link[=prcomp_irlba]{irlba::prcomp_irlba()}} or \code{\link[=prcomp]{base::prcomp()}} should be used for PCA}
}
\value{
Logical indicating if object is an outlier in pca space, or a list if return.pcs=TRUE
}
\description{
Find outliers in the PCA space of a TglowAssay. Works for both
assays and image.data
}
\details{
To run on image data, just specify assay="image.data"|"image.data.trans"|"image.data.norm" which is implemented
as a special case


This does not use existing dimension reductions to avoid issues when running with different QC groups


NA values are removed on a column basis and a warning message is raised. If you want
to handle them differenlty, could make a new assay where the NA's are handled manually

By default, estimates the 0.25 * ncol(assay) PC's assuming this is enough PC's to get to 75%. If this is
not enough a warning is raised, and you can provide pc.max to override this behaviour or set use_irlba
to FALSE to compute all components using \code{\link[=prcomp]{base::prcomp()}}
}
