% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tglow_dataset.r
\name{getImageData,TglowDataset-method}
\alias{getImageData,TglowDataset-method}
\title{Fetch image data or meta data from a tglow object}
\usage{
\S4method{getImageData}{TglowDataset}(object, j, slot = "data", drop = T)
}
\arguments{
\item{object}{TglowDataset}

\item{j}{character with column names from image.meta or image.data to select}

\item{slot}{slot to fetch features fromdata or scale.data}

\item{drop}{should cols be dropped or not}
}
\value{
A data frame with the corresponding columns
}
\description{
Fetch image data or meta data from a tglow object
}
