% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tglow_dataset.r
\name{getImageDataAndFeatures,TglowDataset-method}
\alias{getImageDataAndFeatures,TglowDataset-method}
\title{Get image data and features per cell}
\usage{
\S4method{getImageDataAndFeatures}{TglowDataset}(object, j, assay, slot = "data", drop = F)
}
\arguments{
\item{object}{TglowDataset}

\item{j}{Character with column names from image.meta, image.data@slot or assay@slot to select}

\item{assay}{The assay to select from}

\item{slot}{Slot to fetch features from: "data" or "scale.data"}

\item{drop}{Should cols be dropped or not}
}
\value{
A data frame with the corresponding columns
}
\description{
Select columns from image.data@slot, image.meta, and assay\@slot and return them
as a data.frame.
}
