% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{img_max_project}
\alias{img_max_project}
\title{Max project a EBImage along the 4th axis}
\usage{
img_max_project(img)
}
\arguments{
\item{img}{EBImage}
}
\description{
Max project a EBImage along the 4th axis
}
