% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.r
\name{img_norm}
\alias{img_norm}
\title{Normalize a list of EBImages to a common norm factor}
\usage{
img_norm(images, norm.factors = NULL, q = 1)
}
\arguments{
\item{images}{A list of EBImage objects, with 3 or 4 dimensions. 3d dimension
is assumed to be channel.}

\item{norm.factors}{Value to divide the images by
Defaults to NULL in which case tglow.max.per.channel is called}

\item{q}{Quantile to normalize for, passed to \code{\link{img_max_per_channel}}

importFrom EBImage imageData}
}
\description{
Normalize a list of EBImages to a common norm factor
}
