% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{lm_matrix}
\alias{lm_matrix}
\title{Calculate linear coefficients}
\usage{
lm_matrix(
  response,
  design,
  covariates.dont.use = NULL,
  residuals.only = FALSE,
  return.residuals = FALSE,
  keep.zerocol = FALSE,
  tol = 1e-07,
  calculate_ll = TRUE,
  use.qr = TRUE
)
}
\arguments{
\item{response}{A matrix with response variables}

\item{design}{The common design matrix to regress response againsts}

\item{covariates.dont.use}{Only use if you understand the implications. See details}

\item{residuals.only}{Only return the residual matrix}

\item{return.residuals}{Return residual matrix in the output list. Defaults to T if residual.only = TRUE}

\item{keep.zerocol}{Keep columns in design matrix that sum to zero}

\item{tol}{Minimal eigenvalue to keep columns in the design matrix}

\item{calculate_ll}{Calculate the log likelihood of the model}

\item{use.qr}{Use the qr decomposition instead of the cholesky decomposition on the design matrix. See details}
}
\value{
A list with regression results
}
\description{
Fit a linear mixed model using OLS and find coefficients
}
\details{
This function re-uses the decompostion of the design matrix between different responses, which makes it much faster compared to looped \code{\link[=lm]{lm()}} calls.
It also has extra options such as ignoring covariates when calculating residuals.

\code{use.qr}
If TRUE, the qr decomposition is used to calculate the effectsizes. This avoids a matrix inversion that can be unstable with highly correlated designs.
Singularites are removed based on the provided tol and the absolute diagonal of the QR decomp.
It uses the same basic implementation as \code{\link[=lm]{lm()}}. This is the recommended way

If FALSE, the following is computed. In most cases, it procudes identical results, but there are cases where this is unstable.
Singularities are removed based on the eigenvalue of crossprod(design) and the provided tol.
b <- chol2inv(chol(crossprod(design)))

\code{covariates.dont.use}
Fit the model with these covariates included, but don't use them for calculating residuals (affects r2, se etc)
The covariates for these variables are removed in the output. When specifying more complex models in formula
use the term names, with for instance, interaction terms for example, if it has a form of '~ a + b + c + b:c' and you don't
want to consider the interaction term, add 'b:c'. To remove the intercept, add '(Intercept)'
}
