% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{lm_matrix}
\alias{lm_matrix}
\title{Calculate linear coefficients}
\usage{
lm_matrix(response, design, covariates.dont.use = NULL)
}
\arguments{
\item{response}{A matrix with response variables}

\item{design}{The common design matrix to regress response againsts}

\item{covariates.dont.use}{Fit the covariates, but don't output them or use them for calculating residuals (affects r2 etc)}
}
\value{
A list with regression results
}
\description{
Fit a linear model using OLS and find coefficients
}
\details{
This is more efficient as the b component of the design matrix can be re-used between regressions
b <- chol2inv(chol(crossprod(design)))
}
