% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{lm_matrix}
\alias{lm_matrix}
\title{Calculate linear coefficients}
\usage{
lm_matrix(
  response,
  design,
  covariates.dont.use = NULL,
  residuals.only = FALSE,
  return.residuals = FALSE,
  keep.zerocol = FALSE,
  eigen.tol = 1e-08
)
}
\arguments{
\item{response}{A matrix with response variables}

\item{design}{The common design matrix to regress response againsts}

\item{covariates.dont.use}{Only use if you understand the implications. See detaills}

\item{residuals.only}{Only return the residual matrix}

\item{return.residuals}{Return residual matrix in the output list. Defaults to T if residual.only = TRUE}

\item{keep.zerocol}{Keep columns in design matrix that sum to zero}

\item{eigen.tol}{Minimal eigenvalue to keep columns in the design matrix}
}
\value{
A list with regression results
}
\description{
Fit a linear mixed model using OLS and find coefficients
}
\details{
This is more efficient as the b component of the design matrix can be re-used between regressions
b <- chol2inv(chol(crossprod(design)))

\code{covariates.dont.use}
Fit the model with these covariates included, but don't use them for calculating residuals (affects r2, se etc)
The covariates for these variables are removed in the output. When specifying more complex models in formula
use the term names, with for instance, interaction terms for example, if it has a form of '~ a + b + c + b:c' and you don't
want to consider the interaction term, add 'b:c'. To remove the intercept, add '(Intercept)'
}
