% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{lm_matrix_lrt}
\alias{lm_matrix_lrt}
\title{Calculate LRT on results from lm_matrix}
\usage{
lm_matrix_lrt(res, res.reduced, design, design.null)
}
\arguments{
\item{res}{The results from lm_matrix}

\item{res.reduced}{The results from lm_matrix for the reduced model}

\item{design}{The design matrix for the full model}

\item{design.null}{The design matrix for the reduced model}
}
\description{
Calculates both the chisquared and the f-test based LRT. F test is more accurate for small samples
}
