% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.r
\name{lmm_matrix}
\alias{lmm_matrix}
\title{Calculate linear coefficients using a Linear mixed model}
\usage{
lmm_matrix(
  response,
  design,
  formula,
  formula.null = NULL,
  residuals.only = FALSE,
  return.residuals = FALSE,
  refit = FALSE,
  ...
)
}
\arguments{
\item{response}{A matrix with response variables}

\item{design}{The matrix with predictor variables}

\item{formula}{The latter component of the formula. I.e. \code{~ x + (1|donor)}}

\item{formula.null}{The null formula for a LRT. The latter component of the formula. I.e. \code{~ x + (1|donor)}}

\item{residuals.only}{Only return the residual matrix}

\item{return.residuals}{Return residual matrix in the output list. Defaults to T if residual.only = TRUE}

\item{refit}{Refit the models using MLE during the LRT (anova)}

\item{...}{Remaining parameters passed to \code{\link[lmerTest:lmer]{lmerTest::lmer()}}}
}
\value{
A list with regression results
}
\description{
Fit a linear mixed model using REML / MLE (lme4) and find coefficients
}
