% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{match_objects_xy_nn}
\alias{match_objects_xy_nn}
\title{Match two objects together based on shared images and nearest neighbour in XY}
\usage{
match_objects_xy_nn(
  a,
  b,
  tol = 2,
  mode = "add",
  assay.prefix = "b_",
  meta.prefix = NULL
)
}
\arguments{
\item{a}{A \linkS4class{TglowDataset} used as reference}

\item{b}{A \linkS4class{TglowDataset} used as query (added to reference)}

\item{tol}{Euclidian distance in xy space to consider the NN as valid}

\item{mode}{'add' | 'merge', see details}

\item{assay.prefix}{The prefix to add to assay names when assays in a and b have the same name and mode = merge}

\item{meta.prefix}{The prefix to add to b@meta columns prior to merging, columns with the same name as a in b are dropped}
}
\value{
The object a with the extra data from b
}
\description{
Match objects between two TglowDatasets based on nearest-neighbour in XY coordinate space within shared plate/well/field groups
}
\details{
In mode add all assays from b are added to a as new assays. In mode merge, assays with the same name
are merged together, only new features are added. If the @scale.data slot is missing in one of the
datasets this is set to NULL

NOTE: Image level data is not merged at this time
}
