% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{modul}
\alias{modul}
\alias{modul_std}
\title{As trafo's implementation
https://github.com/cran/trafo/blob/master/R/trafos.R}
\usage{
modul(y, lambda = lambda)

modul_std(y, lambda)
}
\description{
As trafo's implementation
https://github.com/cran/trafo/blob/master/R/trafos.R

Standardized transformation: Modulus
}
