% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{modulus}
\alias{modulus}
\title{Transform vector using modulus transform, finding optimal lambda}
\usage{
modulus(
  object,
  lambda = seq(-2, 2, 1/10),
  plotit = TRUE,
  interp = (plotit && (m < 100)),
  eps = 1/50,
  xlab = expression(lambda),
  ylab = "log-Likelihood",
  ...
)
}
\description{
Modelled after MASS:boxcox() and shares the same basic signature
So see \code{\link[=MASS:boxcox]{MASS:boxcox()}} for help
}
\details{
Implements modulus transformation described here:
https://scales.r-lib.org/reference/boxcox_trans.html
John, J. A., & Draper, N. R. (1980). An alternative family of transformations. Applied Statistics, 190-197. http://www.jstor.org/stable/2986305
}
