% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{skewness}
\alias{skewness}
\alias{kurtosis}
\alias{geometric_mean}
\alias{mode}
\title{Calculate the skewness, kurtosis, geometric mean, mode}
\usage{
skewness(x, na.rm = F)

kurtosis(x, na.rm = F)

geometric_mean(x, na.rm = TRUE)

mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Should NA's be removed}
}
\value{
estimates for skewness, kurtosis respecitvely
}
\description{
Calculate the skewness, kurtosis, geometric mean, mode
}
