% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{objectIds}
\alias{objectIds}
\alias{objectIds<-}
\title{Get and set object Ids}
\usage{
objectIds(object)

objectIds(object) <- value
}
\arguments{
\item{object}{The object to get or set}

\item{value}{If setting using <- the new object ids to assing}
}
\value{
The object with the new Ids
}
\description{
Get and set object Id's on TglowDataset, TglowAssay and TglowReduction
}
