% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{new_384_plate}
\alias{new_384_plate}
\alias{new_96_plate}
\alias{well_to_index}
\title{Utilities for making plate overviews}
\usage{
new_384_plate()

new_96_plate()

well_to_index(well)
}
\description{
Utilities for making plate overviews
}
