% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_hist_dens_grouped}
\alias{plot_hist_dens_grouped}
\title{Plot a grouped histogram}
\usage{
plot_hist_dens_grouped(
  x,
  z,
  xlab,
  main = NULL,
  bins = 100,
  facet = NULL,
  facet.ncol = NULL,
  density = FALSE
)
}
\arguments{
\item{x}{x}

\item{z}{grouping variable}

\item{xlab}{xlab}

\item{main}{title}

\item{bins}{bins}

\item{facet}{Vector with facet variable}

\item{facet.ncol}{Number of columns in facet}

\item{density}{Plot as a density plot instead}
}
\value{
A ggplot2 object
}
\description{
Plot a grouped histogram
}
