% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_img}
\alias{plot_img}
\title{Plot an individual EBImage image with a title}
\usage{
plot_img(
  img,
  main = "img",
  marker.add = T,
  marker.x = NULL,
  marker.y = NULL,
  marker.col = "white",
  marker.size = 8,
  marker.shape = 18
)
}
\arguments{
\item{img}{The image}

\item{main}{title}

\item{marker.add}{Should a marker be added at the center of the image}

\item{marker.x}{x position of marker. If NULL = center}

\item{marker.y}{y position of marker. If NULL = center}

\item{marker.col}{color of the marker}

\item{marker.size}{size of marker}

\item{marker.shape}{shape of marker}
}
\value{
A ggplot2 object
}
\description{
Plot an individual EBImage image with a title
}
