% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_xy_interactive}
\alias{plot_xy_interactive}
\title{Plot an interactive scatterplot with a hover over tooltip or cell image}
\usage{
plot_xy_interactive(
  x,
  y,
  tooltip,
  col = "grey",
  xlab = "X",
  ylab = "Y",
  main = "Interactive plot",
  opts = ggiraph::opts_tooltip(opacity = 1),
  add_func = tglowr::theme_plain,
  labs = NULL,
  labs.add = TRUE,
  labs.size = 5,
  labs.textcol = "white",
  labs.bgcol = "black",
  labs.scalename = "color",
  ...
)
}
\arguments{
\item{x}{The x-axis}

\item{y}{The y-axis}

\item{tooltip}{Text to display in tooltip, can contain HTML, see details}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{main}{Plot title}

\item{opts}{\code{\link[=opts_tooltip]{opts_tooltip()}} opts object to control eastetics}

\item{add_func}{Function that updates the ggplot object, see details}

\item{labs}{dataframe with x, y and label for ident labels to add, autogenerated for factors and character idents under 50 classes}

\item{labs.add}{Should labels be added to the plot}

\item{labs.size}{Point size of text labels}

\item{labs.textcol}{Text color of the labels}

\item{labs.bgcol}{Color of the background behind the text}

\item{labs.scalename}{Name for the colorscale}

\item{...}{Remaining arguments passed to \code{\link[ggiraph:girafe]{ggiraph::girafe()}}}

\item{ident}{Ident to color on}
}
\value{
An interactive ggplot2 object
}
\description{
Plot an interactive scatterplot with a hover over tooltip or cell image
}
\details{
For displaying images with tooltip, specifiy an \if{html}{\out{<img>}} tag with base64
endocded PNG images. To create the tags from a matrix see \code{\link[=img_to_base64png]{img_to_base64png()}}

add_func should be a function that takes a ggplot object, adds things to the ggplot
object and returns the object. Default is \code{\link[=theme_plain]{theme_plain()}}. This can be used for updating
scales or overriding aes
}
