% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_markers}
\alias{plot_markers}
\title{Plot marker analysis results as a bubble plot}
\usage{
plot_markers(
  markers,
  grouping.x = "feature",
  grouping.x2 = NULL,
  feature.clust = NULL,
  annot = NULL,
  size.total = T,
  ylab = "class",
  xlab = "group",
  return.data = F,
  flip.axes = F,
  adjust = "bonferroni",
  alpha = 0.05,
  abs.effect = 0,
  topn = NULL,
  grid.space = "free"
)
}
\arguments{
\item{markers}{A tglow.markers object from find_markers() or find_markers_lmm()}

\item{grouping.x}{Column name in markers or character vector for x-axis grouping. Defualts to "feature" (plot each marker)}

\item{grouping.x2}{Optional second grouping level for faceting}

\item{feature.clust}{Column name for feature clustering, features in same cluster are averaged first}

\item{annot}{Optional annotation data.frame with rownames matching features}

\item{size.total}{If TRUE, calculate percentages using total features, if FALSE use significant only}

\item{ylab}{Label for y-axis}

\item{xlab}{Label for x-axis}

\item{return.data}{If TRUE returns list with plot data and plot, if FALSE returns only plot}

\item{flip.axes}{If TRUE flips x and y axes}

\item{adjust}{Method for p-value adjustment, see ?p.adjust}

\item{alpha}{Significance threshold for adjusted p-values}

\item{abs.effect}{Minimum absolute effect size to consider a marker significant}

\item{topn}{If not NULL, only plot the top n markers per class based on effect size, ignores significance and grouping.x, grouping.x2}

\item{grid.space}{Grid space parameter for faceting, see \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}}
}
\value{
A ggplot2 object or if return.data=TRUE, a list with:
\itemize{
\item data: Data frame with plotting data
\item plot: ggplot2 object
}
}
\description{
Creates a bubble plot visualization of marker analysis results, where dot size
represents percentage of total (or significant) features (or feature groups)
and color represents effect size (or average effect size).
}
\details{
The plot shows markers as points where:
\itemize{
\item Size represents percentage of significant features in each group
\item Color represents effect size (estimate) from statistical test
\item Y-axis shows marker classes
\item X-axis shows groups defined by grouping.x
\item Optional faceting by grouping.x2
}
}
