% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_violin}
\alias{plot_violin}
\title{Plot boxplot with a numerically ordered x and a loess line}
\usage{
plot_violin(
  x,
  y,
  violin = T,
  facet = NULL,
  q = c(0.05, 0.5, 0.95),
  main = NULL,
  xlab = "x",
  ylab = "y"
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{violin}{Should violin plot be added}

\item{facet}{Vector with facet variable}

\item{q}{Quantiles to add as horizontal lines in the density plot}

\item{main}{title}

\item{xlab}{xlab}

\item{ylab}{ylab}
}
\value{
A ggplot2 object
}
\description{
Plot boxplot with a numerically ordered x and a loess line
}
