% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{read_cellprofiler_dir}
\alias{read_cellprofiler_dir}
\title{Read a cellprofiler fileset directory tree}
\usage{
read_cellprofiler_dir(
  path,
  pattern,
  type,
  n = NULL,
  skip.orl = TRUE,
  verbose = F,
  col.object = "cell_ObjectNumber_Global",
  col.meta.img.id = "ImageNumber_Global",
  ...
)
}
\arguments{
\item{path}{path to tglow output dir}

\item{pattern}{The pattern that uniquely identfies a fileset. Use '.zip' for type 'B'
and the '_Image.txt' or '_Experiment.txt'(or however you exported the image data) for type 'A'
for type A}

\item{type}{Must be 'A' or 'B'. See details}

\item{n}{Read a subset of filesets. If integer, only that fileset is read, otherwise specify indices to read}

\item{skip.orl}{Skip reading of object relationships, as this can get quite large with many children and is not used}

\item{verbose}{Should I be chatty?}

\item{col.object}{The collumn name in the features which contains the per object object identifier. See details}

\item{col.meta.img.id}{The collumn name in the image level data which contains the image id. See details}

\item{...}{Remaining parameters passed to \code{\link{read_cellprofiler_fileset_a}} or \code{\link{read_cellprofiler_fileset_b}}}
}
\description{
Read a cellprofiler fileset directory tree organized into a \if{html}{\out{<plate>}}/\if{html}{\out{<well>}}/\if{html}{\out{<field>}}.fileset
structure
}
\details{
\code{type}
Type A: _cell.txt, _Image.txt, _Experiment.txt and _Object relationships.txt
See \code{\link{read_cellprofiler_fileset_a}} for detaills

Type B: \if{html}{\out{<plate>}}_\if{html}{\out{<well>}}.zip with individual files for each child object. Main object is assumed to be _cells
See \code{\link{read_cellprofiler_fileset_b}} for detaills

\code{col.object} and \code{col.meta.img.id}

See \code{\link{add_global_ids}} for details on how globally unique Id's are assigned in the case you need to set
\code{col.object} or `col.meta.img.id``.
}
