% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{read_cellprofiler_dir}
\alias{read_cellprofiler_dir}
\title{Read a cellprofiler fileset directory tree}
\usage{
read_cellprofiler_dir(
  path,
  pattern,
  type,
  n = NULL,
  skip.orl = TRUE,
  verbose = F,
  col.object = "cell_ObjectNumber_Global",
  col.meta.img.id = "ImageNumber_Global",
  ...
)
}
\arguments{
\item{path}{path to tglow output dir}

\item{pattern}{The pattern that uniquely identfies a well. Use '.zip' for type 'B' and the '_experiment.tsv' for type 'A'
for type A}

\item{type}{Must be 'A' or 'B'. See details}

\item{n}{Read a subset of filesets. If integer, only that fileset is read, otherwise specify indices to read}

\item{skip.orl}{Skip reading of object relationships, as this can get quite large with many children and is not used}

\item{verbose}{Should I be chatty?}

\item{col.object}{The collumn name in the features which contains the per object object identifier}

\item{col.meta.img.id}{The collumn name in the image level data which contains the image id}

\item{...}{Remaining parameters passed to \code{\link{read_cellprofiler_fileset_a}} or \code{\link{read_cellprofiler_fileset_b}}}
}
\description{
Matches images based on \if{html}{\out{<plate>}}, \if{html}{\out{<well>}}, \if{html}{\out{<field>}}
}
\details{
Type A: _cells.tsv, _image.tsv, _experiment.tsv and _objectRelations.tsv
See r\code{\link{read_cellprofiler_fileset_a}} for detaills

Type B: \if{html}{\out{<plate>}}_\if{html}{\out{<well>}}.zip with individual files for each child object. Main object is assumed to be _cells
See r\code{\link{read_cellprofiler_fileset_b}} for detaills
}
