% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{read_cellprofiler_fileset_a}
\alias{read_cellprofiler_fileset_a}
\title{Read a cell level fileset type A}
\usage{
read_cellprofiler_fileset_a(
  prefix,
  return.feature.meta = F,
  add.global.id = T,
  pat.img = "_image.tsv",
  pat.cells = "_cells.tsv",
  pat.orl = "_objectRelationships.tsv"
)
}
\value{
list with data frames:

- cells (cell level features)

- meta (image level features)

- objectRelations

- features [optional]


Output is NULL if no cells are detected.
}
\description{
Reads a CellProfiler fileset into a list.
Type A: Assumes all features are in a single _cells.tsv / _cells.tsv
and all features are matched
}
