% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{read_cellprofiler_fileset_a}
\alias{read_cellprofiler_fileset_a}
\title{Read a cell level fileset type A}
\usage{
read_cellprofiler_fileset_a(
  prefix,
  return.feature.meta = F,
  add.global.id = T,
  pat.img = "_image.tsv",
  pat.cells = "_cells.tsv",
  pat.orl = "_objectRelationships.tsv",
  skip.orl = FALSE
)
}
\arguments{
\item{prefix}{Path prefix to fileset}

\item{return.feature.meta}{Should the dataframe with feature metadata be added}

\item{add.global.id}{Should extra id columns be added that are globally unique}

\item{pat.img}{The suffix pattern to identify the image level data}

\item{pat.cells}{The suffix pattern to identify the cell level data}

\item{pat.orl}{The suffix pattern to identify object relationships}

\item{skip.orl}{Should object relationships be read (not used, can be quite large)}
}
\value{
list with data frames:
\itemize{
\item cells (cell level features)
\item meta (image level features)
\item objectRelations
\item features \link{optional}
}

Output is NULL if no cells are detected
}
\description{
Reads a CellProfiler fileset into a list
Type A: Assumes all features are in a single _cells.tsv / _cells.tsv
and all features are matched
}
