% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{read_cellprofiler_fileset_b}
\alias{read_cellprofiler_fileset_b}
\title{Read a cell level fileset type B}
\usage{
read_cellprofiler_fileset_b(
  prefix,
  return.feature.meta = F,
  add.global.id = T,
  merging.strategy = "mean",
  parent.col = "Parent_cell",
  pat.exp = ".*Experiment.txt",
  pat.img = ".*Image.txt",
  pat.cells = ".*cell.txt",
  pat.orl = ".*Object relationships.txt",
  pat.others = "^.*_([a-zA-Z]+\\\\d*).txt$",
  na.rm = F,
  skip.orl = F,
  verbose = F,
  fileset.id = NULL
)
}
\arguments{
\item{prefix}{Path prefix to fileset}

\item{return.feature.meta}{Should the dataframe with feature metadata be added}

\item{add.global.id}{Should extra id columns be added that are globally unique}

\item{merging.strategy}{How to consolidate 1:many relationships between cell: children. Accepted values: 'mean', 'none'}

\item{pat.exp}{The suffix pattern to identify the exeperiment file}

\item{pat.img}{The suffix pattern to identify the image level data}

\item{pat.cells}{The suffix pattern to identify the cell level data}

\item{pat.orl}{The suffix pattern to identify object relationships}

\item{pat.others}{The pattern to use to extract the child object names from the filename. First regex group is used as object name}

\item{na.rm}{Should NA's be removed when applying merging.strategy}

\item{skip.orl}{Should object relationships be read (not used, can be quite large)}

\item{verbose}{Should I be chatty?}

\item{fileset.id}{The global fileset id to add if add.global.id = TRUE}
}
\value{
list with data frames:
\itemize{
\item cells (cell level features)
\item meta (image level features)
\item objectRelations
\item features (optional)
Output is NULL if no cells are detected
}
}
\description{
Reads a .zip file with cellprofiler features, each file other then
_Image, _Experiment and _Object Relationships are assumed to be an object
Will match objects on order with an appropriate matching strategy
}
