% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{scale_assay}
\alias{scale_assay}
\title{Scale a TglowAssay}
\usage{
scale_assay(assay, grouping = NULL, reference.group = NULL, ...)
}
\arguments{
\item{assay}{\linkS4class{TglowAssay}}

\item{grouping}{Vector with a grouping variable of length nrow(assay)}

\item{reference.group}{Scale not to the vector mean/median, sd/mad but to the objects indiciated here}

\item{...}{Arguments passed to \code{\link{fast_colscale}}. Strongly reccomend looking at these!}
}
\value{
The asssay with the scale.data slot populated
}
\description{
Scale the data slot in a tglow assay and override the scale.data slot
Defaults to z-score, but can also return modified z-score when providing scale.method
}
\details{
\code{grouping}

When this is provided, scaling is done within each unique value of this vector

\code{reference.group}

When this is provided, the data is scaled to the mean/median, sd/mad of only
the samples set to TRUE in this vector. It much be a logical vector indicating
if the object is a reference object. This is usefull to scaling imaging features
relative to a control sample.

\code{grouping} and \code{reference.group}

All selection vectors are assumed to have the same row order as assay
groupung and reference.group interact, so for example if you want to scale to
the controls on a plate you would set grouping to getDataByObject(object, "plate")
and reference.group to getDataByObject(object, "is_control")
}
