% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{scale_dataset}
\alias{scale_dataset}
\title{Scale a tglow dataset}
\usage{
scale_dataset(dataset, assay = NULL, grouping = NULL, ...)
}
\arguments{
\item{dataset}{\linkS4class{TglowDataset}}

\item{assay}{Character string indicating assay, if NULL all assays are scaled.}

\item{grouping}{Colname on dataset or Vector with a grouping variable of length nrow(assay)}

\item{...}{Remaining options passed to \code{\link[=scale_assay]{scale_assay()}}}
}
\value{
The \linkS4class{TglowDataset} with assay scale.data slots populated
}
\description{
For TglowDataset, scales the provided assay, or all assays when assay is null
}
