% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{aggregate_tglow_matrix}
\alias{aggregate_tglow_matrix}
\alias{aggregate_assay}
\alias{aggregate_metadata}
\alias{aggregate_by_imagecol}
\title{Aggregate tglow objects}
\usage{
aggregate_tglow_matrix(
  matrix,
  grouping,
  method = "mean",
  group.order = NULL,
  na.rm = TRUE
)

aggregate_assay(assay, grouping, method, group.order = NULL, na.rm = TRUE)

aggregate_metadata(
  data,
  grouping,
  method,
  group.order = NULL,
  na.rm = TRUE,
  drop.multival.string = TRUE,
  drop.na.col = TRUE,
  sep = "; "
)

aggregate_by_imagecol(
  object,
  grouping,
  method,
  group.order = NULL,
  na.rm = TRUE,
  drop.multival.string = TRUE,
  drop.na.col = TRUE,
  sep = "; "
)
}
\arguments{
\item{grouping}{Grouping variable of nrow input. For aggregate_by_imagecol, grouping must be a column name on image.meta}

\item{method}{Aggregation method: "mean", "median", "sum"}

\item{group.order}{The row order of output}

\item{na.rm}{Should NA's be removed}

\item{drop.multival.string}{When concatenating character metadata, should strings with multiple values be set to NA. This is reccomended for large datasets.}

\item{drop.na.col}{If after aggregating, a column is fully NA, should it be dropped?}

\item{sep}{Seperator string when concatenating multiple strings}

\item{input}{Input: either TglowMatrix, TglowAssay, data.frame or TglowDataset}
}
\value{
A logical where TRUE should be kept and FALSE values should be removed.
}
\description{
Aggregate tglow objects using a grouping variable
}
\details{
Aggregates objects using mean, median or sum based on a grouping variable.
Strings are returned if there is a single unique value, otherwise they are dropped by default.
Set drop.multival.string = FALSE to retain multival strings, which are pasted together using sep.
drop.na.col removes any columns if drop.multival.string = TRUE and the result has only NA's left.

Both assays and image.data, image.data.trans and image.data.norm are aggregated.
}
