% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.r
\name{TglowDatasetFromMatrices}
\alias{TglowDatasetFromMatrices}
\alias{TglowDatasetFromObjectMatrix}
\title{Create a new TglowDataset}
\usage{
TglowDatasetFromMatrices(
  objects,
  images,
  image.ids,
  object.meta = NULL,
  image.meta = NULL,
  assay.out = "raw"
)

TglowDatasetFromObjectMatrix(objects, object.meta = NULL, assay.out = "raw")
}
\arguments{
\item{objects}{The object level data matrix, must be a matrix with row and column names set}

\item{image.ids}{A character vector of length nrow(objects) indicating which row of the image matrix corresponds to each object}

\item{object.meta}{Optional data.frame with object level metadata. Defaults to a minimal data.frame with row ids}

\item{image.meta}{A data.frame with metadata, one row per image. Defaults to data.frame(id=rownames(images))}

\item{assay.out}{The assay name to store objects under}

\item{image}{The image level data matrix, must be a matrix, and be numeric}
}
\value{
A populated TglowDataset
}
\description{
Create a new TglowDataset object from a matrix with image level data and a matrix of object level data
}
\details{
Only the data slot on assay is populated with the values in \code{objects} if this is not what you want you can
always re-assign data to the scale.data slot afterwards by \code{dataset$raw@scale.data <- dataset$raw@data}.
Just to note, the assumption is made that the scale.data slot is actually scaled to mean=0 variance=1 when
running PCA's etc, so overriding with non-scaled data may change your results.
}
