% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{tglow_dimplot}
\alias{tglow_dimplot}
\title{Plot a reduction}
\usage{
tglow_dimplot(
  object,
  reduction,
  ident = NULL,
  assay = NULL,
  slot = NULL,
  downsample = NULL,
  log.ident = FALSE,
  axis.x = 1,
  axis.y = 2,
  xlab = NULL,
  ylab = NULL,
  no.colscale = FALSE,
  ...
)
}
\arguments{
\item{reduction}{A name of a reduction on dataset}

\item{ident}{The item to use for coloring points, passed to \code{\link{getDataByObject}}}

\item{assay}{The assay to use for coloring, passed to \code{\link{getDataByObject}}}

\item{slot}{The slot to use for coloring, passed to \code{\link{getDataByObject}} Can be "data" or "scale.data"}

\item{downsample}{Should downsampling be applied prior to plot. NA's are removed first}

\item{log.ident}{Should color vector be log2 transformed}

\item{axis.x}{Which dimension from reduction to plot in x}

\item{axis.y}{Which dimension from reduction to plot in y}

\item{xlab}{Override xlab}

\item{ylab}{Override ylab}

\item{no.colscale}{Skip adding colorscale}

\item{...}{Remaining arguments passed to \code{\link{plot_xy}}}

\item{dataset}{A \linkS4class{TglowDataset}}
}
\description{
Plot a reduction on a \linkS4class{TglowDataset}
}
