% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{tglow_dimplot_interactive}
\alias{tglow_dimplot_interactive}
\title{Plot an interactive scatterplot with a hover over tooltip or cell image}
\usage{
tglow_dimplot_interactive(
  dataset,
  ident,
  reduction = NULL,
  images = NULL,
  tooltip = NULL,
  dim.x = 1,
  dim.y = 2,
  image.size = 250,
  xlab = NULL,
  ylab = NULL,
  x = NULL,
  y = NULL,
  assay = NULL,
  slot = NULL,
  ...
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{ident}{Column to color on}

\item{reduction}{Reduction to use}

\item{images}{List of RGB images, assumed to be in the same order as dataset}

\item{tooltip}{Character vector of length objects of tooltips to add}

\item{dim.x}{The x-axis to use}

\item{dim.y}{The y-axis to use}

\item{image.size}{The size to render the image in the interactive plot}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{x}{Override x}

\item{y}{Override y}

\item{...}{remaining params passed to \code{\link[=plot_xy_interactive]{plot_xy_interactive()}}}
}
\description{
Plot an interactive scatterplot with a hover over tooltip or cell image
}
