% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cellprofiler.r
\name{tglow_feature_map_cellprofiler}
\alias{tglow_feature_map_cellprofiler}
\title{Default feature map for CellProfiler parquet output}
\usage{
tglow_feature_map_cellprofiler()
}
\value{
A \linkS4class{TglowFeatureMap}
}
\description{
Returns the \linkS4class{TglowFeatureMap} matching the output of \code{\link{read_cellprofiler_parquet}}.
x/y/z, plate and well are taken from \verb{@meta}, field from \verb{@image.meta}.
}
