% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{f.min}
\alias{f.min}
\alias{f.min.sum}
\alias{f.max}
\alias{f.max.sum}
\alias{f.na}
\alias{f.na.multicol}
\alias{f.near.zero.var}
\alias{f.near.zero.var.sum}
\alias{f.zero.var}
\alias{f.zero.var.sum}
\alias{f.coef.var}
\alias{f.coef.var.sum}
\alias{f.unique.val}
\alias{f.unique.val.sum}
\alias{f.inf.median}
\alias{f.inf.median.sum}
\alias{f.inf}
\alias{f.inf.mutlicol}
\alias{f.mod.z}
\alias{Met}
\alias{f.mod.z.perc}
\alias{f.sum}
\title{Filters to apply to TglowDatasets}
\usage{
f.min(vec, thresh, grouping = NULL)

f.min.sum(...)

f.max(vec, thresh, grouping = NULL)

f.max.sum(...)

f.na(vec, thresh, grouping = NULL)

f.na.multicol(vec, thresh, grouping)

f.near.zero.var(vec, thresh = NULL)

f.near.zero.var.sum(...)

f.zero.var(vec, thresh = 0)

f.zero.var.sum(...)

f.coef.var(vec, thresh)

f.coef.var.sum(...)

f.unique.val(vec, thresh = NULL)

f.unique.val.sum(...)

f.inf.median(vec, thresh = NULL)

f.inf.median.sum(...)

f.inf(vec, thresh = NULL, grouping = NULL)

f.inf.mutlicol(vec, thresh, grouping)

f.mod.z(vec, thresh, grouping = NULL, absolute = T, method = "mod.z")

Met(...)

f.mod.z.perc(vec, thresh, thresh2, grouping = NULL)

f.sum(vec, thresh, grouping, func)
}
\arguments{
\item{vec}{Input vector or value}

\item{thresh}{Filter threshold value}

\item{grouping}{Optional grouping for filters applied on a subgroup of vec, not all filters respect this!}
}
\value{
A logical where TRUE should be kept and FALSE values should be removed.
}
\description{
Filters come in 2 variants, the base function, which is applied over a vector or value.
The base then has a <filter>.sum function which sums over multiple vectors or values.
}
