% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{filter_min}
\alias{filter_min}
\alias{filter_min_sum}
\alias{filter_max}
\alias{filter_max_sum}
\alias{filter_na}
\alias{filter_na_multicol}
\alias{filter_near_zero_var}
\alias{filter_near_zero_var_sum}
\alias{filter_zero_var}
\alias{filter_zero_var_sum}
\alias{filter_coef_var}
\alias{filter_coef_var_sum}
\alias{filter_unique_val}
\alias{filter_unique_val_sum}
\alias{filter_inf_median}
\alias{filter_inf_median_sum}
\alias{filter_inf}
\alias{filter_inf_mutlicol}
\alias{filter_mod_z}
\alias{Met}
\alias{filter_mod_z_perc}
\alias{filter_sum}
\title{Filters to apply to TglowDatasets}
\usage{
filter_min(vec, thresh, grouping = NULL)

filter_min_sum(...)

filter_max(vec, thresh, grouping = NULL)

filter_max_sum(...)

filter_na(vec, thresh, grouping = NULL)

filter_na_multicol(vec, thresh, grouping)

filter_near_zero_var(vec, thresh = NULL)

filter_near_zero_var_sum(...)

filter_zero_var(vec, thresh = 0)

filter_zero_var_sum(...)

filter_coef_var(vec, thresh)

filter_coef_var_sum(...)

filter_unique_val(vec, thresh = NULL)

filter_unique_val_sum(...)

filter_inf_median(vec, thresh = NULL)

filter_inf_median_sum(...)

filter_inf(vec, thresh = NULL, grouping = NULL)

filter_inf_mutlicol(vec, thresh, grouping)

filter_mod_z(vec, thresh, grouping = NULL, absolute = T, method = "mod.z")

Met(...)

filter_mod_z_perc(vec, thresh, thresh2, grouping = NULL)

filter_sum(vec, thresh, grouping, func)
}
\arguments{
\item{vec}{Input vector or value}

\item{thresh}{Filter threshold value}

\item{grouping}{Optional grouping for filters applied on a subgroup of vec, not all filters respect this!}
}
\value{
A logical where TRUE should be kept and FALSE values should be removed
}
\description{
Filters come in 2 variants, the base function, which is applied over a vector or value
The base then has a <filter>.sum function which sums over multiple vectors or values
}
\details{
Uses Rfast whose variance calculations are not exactly zero, so set the zero threshold to 1e-10 by default
}
