% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.r
\name{tglow_filters_from_table}
\alias{tglow_filters_from_table}
\title{Create TglowFilters from a filter table}
\usage{
tglow_filters_from_table(
  filter.table,
  name.col = 1,
  col.col = 2,
  func.col = 3,
  thresh.col = 4,
  trans.col = NULL,
  active.col = NULL
)
}
\arguments{
\item{filter.table}{The table with filters}

\item{name.col}{Column id with filter name}

\item{col.col}{Column id with collumn pattern to apply filter too}

\item{func.col}{Column id with name of filter functions}

\item{thresh.col}{Column id with col with threshold}

\item{trans.col}{Column id with col indicating if transpose is applied before running func}

\item{active.col}{Column id with boolean setting filter to active or inactive}
}
\value{
A list of \linkS4class{TglowFilter}
}
\description{
See  \linkS4class{TglowFilter} for detaills
}
