% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{tglow_plot_execution_time}
\alias{tglow_plot_execution_time}
\title{Chart of average execution times}
\usage{
tglow_plot_execution_time(object)
}
\arguments{
\item{object}{A TglowDataset}
}
\value{
ggplot object containing the plot
}
\description{
Plot the execution time of CellProfiler modules
}
