% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{tglow_plot_location_hex}
\alias{tglow_plot_location_hex}
\title{Plot a hexbin colored on a 3rd variable}
\usage{
tglow_plot_location_hex(
  dataset,
  assay,
  slot,
  feature.z,
  feature.x = "nucl_AreaShape_Center_X",
  feature.y = "nucl_AreaShape_Center_Y",
  xlab = "X centroid",
  ylab = "Y centroid",
  zlab = "Z",
  main = NULL,
  bins = 30,
  scale.to.well.mean = F,
  scale.z = F,
  trim.outliers.z = F,
  trim.outliers.z.thresh = 3.5,
  bins.mincount = 0
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{assay}{The assay to use}

\item{slot}{The slot to use for calculating filters, defaults to "data". Can be "data" or "scale.data"}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{zlab}{Scale title}

\item{main}{title}

\item{bins}{bins}

\item{scale.to.well.mean}{Divide z by the mean (fold change compared to mean)}

\item{scale.z}{Scale z to mean 0 variance 1}

\item{trim.outliers.z}{Remove outliers in z prior to plotting using modified z-score}

\item{trim.outliers.z.thresh}{Modified z-score hreshold to consider an outlier}

\item{bins.mincount}{Minimum number of objects in a bin to render it}
}
\value{
A ggplot2 object
}
\description{
Plot a hexbin colored on a 3rd variable
}
