% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_image.r
\name{tglow_read_imgs}
\alias{tglow_read_imgs}
\title{Read images for objects in a TglowDataset from the tglow-pipeline .h5 format}
\usage{
tglow_read_imgs(
  dataset,
  objects,
  index.col,
  out.size = NULL,
  path = NULL,
  path.col = NULL
)
}
\arguments{
\item{dataset}{A \linkS4class{TglowDataset}}

\item{objects}{A selection vector for which objects to read images for}

\item{index.col}{Column on dataset@meta which has the h5 groups (object ids)}

\item{out.size}{Pad the images to a constant square size setting 0. NULL = no padding.}

\item{path}{The root path for the .h5 files. 'root path'/plate/row/col/field.h5}

\item{path.col}{Column in dataset@meta with the paths to .h5 file. Overrides path if not NULL}
}
\value{
A list with EBImages and channel names
}
\description{
Read image data from a .h5 file that is orgnaized as such:
/plate/row/col/field.h5 where each h5 group represents a cells image
}
