% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{TglowFeatureLocation-class}
\alias{TglowFeatureLocation-class}
\title{TglowFeatureMap}
\description{
TglowFeatureMap
}
\section{Slots}{

\describe{
\item{\code{feature}}{Character with the feature position}

\item{\code{assay}}{to grab feature from, or NULL if metadata feature}

\item{\code{slot}}{to grab feature from, or NULL if metadata feature}
}}

