% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{TglowReduction-class}
\alias{TglowReduction-class}
\title{TglowReduction}
\description{
TglowReduction
}
\section{Slots}{

\describe{
\item{\code{x}}{Slot to store PCA / UMAP coordinates}

\item{\code{sdev}}{optional standard deviations}

\item{\code{sdev_total}}{optional sum of sdev}

\item{\code{object}}{Optional PCA / UMAP object}
}}

