% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{theme_plain}
\alias{theme_plain}
\title{Simple plotting theme wrapper for ggplot objects}
\usage{
theme_plain(p, base_size = 11, base_family = "ArialMT", legend = TRUE)
}
\arguments{
\item{p}{A ggplot object}

\item{base_size}{Text size}

\item{base_family}{Text font}

\item{legend}{Keep the legend}
}
\description{
Simple plotting theme wrapper for ggplot objects
}
